@@ -1,27 +1,31 @@
 from __future__ import annotations
 
 import inspect
-import warnings
 from collections.abc import Mapping
 from dataclasses import dataclass
 from functools import partial, singledispatch, wraps
 from types import MappingProxyType
-from typing import TYPE_CHECKING, Generic, TypeVar
+from typing import TYPE_CHECKING
+
+import numpy as np
 
 from anndata._io.utils import report_read_key_on_error, report_write_key_on_error
+from anndata._settings import settings
 from anndata._types import Read, ReadLazy, _ReadInternal, _ReadLazyInternal
-from anndata.compat import DaskArray, ZarrGroup, _read_attr, is_zarr_v2
+from anndata.compat import DaskArray, ZarrGroup, _read_attr, has_xp
+
+from ...utils import warn
 
 if TYPE_CHECKING:
     from collections.abc import Callable, Generator, Iterable
     from typing import Any
 
     from anndata._types import (
-        GroupStorageType,
         ReadCallback,
         StorageType,
         Write,
         WriteCallback,
+        _GroupStorageType,
         _WriteInternal,
     )
     from anndata.experimental.backed._lazy_arrays import CategoricalArray, MaskedArray
@@ -29,9 +33,22 @@ if TYPE_CHECKING:
 
     from ..._core.xarray import Dataset2D
 
-    T = TypeVar("T")
-    W = TypeVar("W", bound=_WriteInternal)
-    LazyDataStructures = DaskArray | Dataset2D | CategoricalArray | MaskedArray
+    type LazyDataStructures = DaskArray | Dataset2D | CategoricalArray | MaskedArray
+
+
+def to_writeable(x):
+    # Convert non-numpy arrays to dlpack
+    if has_xp(x) and not (isinstance(x, np.ndarray) or np.isscalar(x)):
+        return np.from_dlpack(x)
+    return x
+
+
+def normalize_nested(obj):
+    if isinstance(obj, dict):
+        return {k: normalize_nested(v) for k, v in obj.items()}
+    if isinstance(obj, list | tuple):
+        return type(obj)(normalize_nested(v) for v in obj)
+    return to_writeable(obj)
 
 
 # TODO: This probably should be replaced by a hashable Mapping due to conversion b/w "_" and "-"
@@ -69,10 +86,10 @@ class IORegistryError(Exception):
         return cls(msg)
 
 
-def write_spec(spec: IOSpec):
+def write_spec[W: _WriteInternal](spec: IOSpec) -> Callable[[W], W]:
     def decorator(func: W) -> W:
         @wraps(func)
-        def wrapper(g: GroupStorageType, k: str, *args, **kwargs):
+        def wrapper(g: _GroupStorageType, k: str, *args, **kwargs) -> None:
             result = func(g, k, *args, **kwargs)
             g[k].attrs.setdefault("encoding-type", spec.encoding_type)
             g[k].attrs.setdefault("encoding-version", spec.encoding_version)
@@ -83,20 +100,19 @@ def write_spec(spec: IOSpec):
     return decorator
 
 
-_R = TypeVar("_R", _ReadInternal, _ReadLazyInternal)
-R = TypeVar("R", Read, ReadLazy)
+class IORegistry[RI: (_ReadInternal, _ReadLazyInternal), R: (Read, ReadLazy)]:
+    read: dict[tuple[type, IOSpec, frozenset[str]], RI]
+    read_partial: dict[tuple[type, IOSpec, frozenset[str]], Callable]
+    write: dict[tuple[type, type | tuple[type, str], frozenset[str]], _WriteInternal]
+    write_specs: dict[type | tuple[type, str] | tuple[type, type], IOSpec]
 
+    def __init__(self) -> None:
+        self.read = {}
+        self.read_partial = {}
+        self.write = {}
+        self.write_specs = {}
 
-class IORegistry(Generic[_R, R]):
-    def __init__(self):
-        self.read: dict[tuple[type, IOSpec, frozenset[str]], _R] = {}
-        self.read_partial: dict[tuple[type, IOSpec, frozenset[str]], Callable] = {}
-        self.write: dict[
-            tuple[type, type | tuple[type, str], frozenset[str]], _WriteInternal
-        ] = {}
-        self.write_specs: dict[type | tuple[type, str] | tuple[type, type], IOSpec] = {}
-
-    def register_write(
+    def register_write[T](
         self,
         dest_type: type,
         src_type: type | tuple[type, str],
@@ -118,7 +134,7 @@ class IORegistry(Generic[_R, R]):
         else:
             self.write_specs[src_type] = spec
 
-        def _register(func):
+        def _register(func: _WriteInternal[T]) -> _WriteInternal[T]:
             self.write[(dest_type, src_type, modifiers)] = write_spec(spec)(func)
             return func
 
@@ -154,11 +170,11 @@ class IORegistry(Generic[_R, R]):
         src_type: type,
         spec: IOSpec | Mapping[str, str],
         modifiers: Iterable[str] = frozenset(),
-    ) -> Callable[[_R], _R]:
+    ) -> Callable[[RI], RI]:
         spec = proc_spec(spec)
         modifiers = frozenset(modifiers)
 
-        def _register(func):
+        def _register(func: RI) -> RI:
             self.read[(src_type, spec, modifiers)] = func
             return func
 
@@ -240,12 +256,9 @@ def proc_spec_mapping(spec: Mapping[str, str]) -> IOSpec:
 def get_spec(
     elem: StorageType,
 ) -> IOSpec:
-    return proc_spec(
-        {
-            k: _read_attr(elem.attrs, k, "")
-            for k in ["encoding-type", "encoding-version"]
-        }
-    )
+    return proc_spec({
+        k: _read_attr(elem.attrs, k, "") for k in ["encoding-type", "encoding-version"]
+    })
 
 
 def _iter_patterns(
@@ -305,7 +318,7 @@ class LazyReader(Reader):
         )
         if self.callback is not None:
             msg = "Dask reading does not use a callback. Ignoring callback."
-            warnings.warn(msg, stacklevel=2)
+            warn(msg, UserWarning)
         read_params = inspect.signature(read_func).parameters
         for kwarg in kwargs:
             if kwarg not in read_params:
@@ -338,7 +351,7 @@ class Writer:
     @report_write_key_on_error
     def write_elem(
         self,
-        store: GroupStorageType,
+        store: _GroupStorageType,
         k: str,
         elem: RWAble,
         *,
@@ -349,10 +362,17 @@ class Writer:
 
         import h5py
 
+        from anndata._io.zarr import is_group_consolidated
+
         # we allow stores to have a prefix like /uns which are then written to with keys like /uns/foo
-        if "/" in k.split(store.name)[-1][1:]:
-            msg = "Forward slashes are not allowed in keys."
-            raise ValueError(msg)
+        is_zarr_group = isinstance(store, ZarrGroup)
+        if "/" in k.rsplit(store.name, maxsplit=1)[-1][1:]:
+            if is_zarr_group or settings.disallow_forward_slash_in_h5ad:
+                msg = f"Forward slashes are not allowed in keys in {type(store)}"
+                raise ValueError(msg)
+            else:
+                msg = "Forward slashes will be disallowed in h5 stores in the next minor release"
+                warn(msg, FutureWarning)
 
         if isinstance(store, h5py.File):
             store = store["/"]
@@ -360,24 +380,14 @@ class Writer:
         dest_type = type(store)
 
         # Normalize k to absolute path
-        if (
-            is_zarr_v2_store := (
-                (is_zarr_store := isinstance(store, ZarrGroup)) and is_zarr_v2()
-            )
-        ) or (isinstance(store, h5py.Group) and not PurePosixPath(k).is_absolute()):
+        if isinstance(store, h5py.Group) and not PurePosixPath(k).is_absolute():
             k = str(PurePosixPath(store.name) / k)
-        is_consolidated = False
-        if is_zarr_v2_store:
-            from zarr.storage import ConsolidatedMetadataStore
-
-            is_consolidated = isinstance(store.store, ConsolidatedMetadataStore)
-        elif is_zarr_store:
-            is_consolidated = store.metadata.consolidated_metadata is not None
+        is_consolidated = is_group_consolidated(store) if is_zarr_group else False
         if is_consolidated:
             msg = "Cannot overwrite/edit a store with consolidated metadata"
             raise ValueError(msg)
         if k == "/":
-            if isinstance(store, ZarrGroup) and not is_zarr_v2():
+            if isinstance(store, ZarrGroup):
                 from zarr.core.sync import sync
 
                 sync(store.store.clear())
@@ -385,6 +395,9 @@ class Writer:
                 store.clear()
         elif k in store:
             del store[k]
+
+        # Normalize array-API (e.g., JAX/CuPy) even if not AnnData
+        elem = normalize_nested(elem)
 
         write_func = self.find_write_func(dest_type, elem, modifiers)
 
@@ -495,7 +508,7 @@ def read_elem_lazy(
 
 
 def write_elem(
-    store: GroupStorageType,
+    store: _GroupStorageType,
     k: str,
     elem: RWAble,
     *,
