@@ -3,23 +3,21 @@ from __future__ import annotations
 import inspect
 import os
 import textwrap
-import warnings
 from collections.abc import Iterable
 from contextlib import contextmanager
 from dataclasses import dataclass, field, fields
 from enum import Enum
 from functools import partial
 from inspect import Parameter, signature
-from types import GenericAlias
-from typing import TYPE_CHECKING, Generic, NamedTuple, TypeVar, cast
+from types import GenericAlias, NoneType
+from typing import TYPE_CHECKING, NamedTuple, cast
 
-from .compat import is_zarr_v2, old_positionals
+from ._warnings import warn
+from .compat import old_positionals
 
 if TYPE_CHECKING:
     from collections.abc import Callable, Sequence
-    from typing import Any, TypeGuard
-
-T = TypeVar("T")
+    from typing import Any, Self, TypeGuard
 
 
 class DeprecatedOption(NamedTuple):
@@ -51,17 +49,17 @@ def describe(self: RegisteredOption, *, as_rst: bool = False) -> str:
     return textwrap.dedent(doc)
 
 
-class RegisteredOption(NamedTuple, Generic[T]):
+class RegisteredOption[T](NamedTuple):
     option: str
     default_value: T
     description: str
-    validate: Callable[[T], None]
+    validate: Callable[[T, SettingsManager], None]
     type: object
 
     describe = describe
 
 
-def check_and_get_environ_var(
+def check_and_get_environ_var[T](
     key: str,
     default_value: str,
     allowed_values: Sequence[str] | None = None,
@@ -93,7 +91,7 @@ def check_and_get_environ_var(
             f"Value {environ_value_or_default_value!r} is not in allowed {allowed_values} for environment variable {key}. "
             f"Default {default_value} will be used."
         )
-        warnings.warn(msg, UserWarning, stacklevel=3)
+        warn(msg, UserWarning)
         environ_value_or_default_value = default_value
     return (
         cast(environ_value_or_default_value)
@@ -102,7 +100,7 @@ def check_and_get_environ_var(
     )
 
 
-def check_and_get_bool(option, default_value):
+def check_and_get_bool(option: str, default_value: bool) -> bool:  # noqa: FBT001
     return check_and_get_environ_var(
         f"ANNDATA_{option.upper()}",
         str(int(default_value)),
@@ -111,12 +109,18 @@ def check_and_get_bool(option, default_value):
     )
 
 
-def check_and_get_int(option, default_value):
+def check_and_get_bool_or_none(option: str, default_value: bool | None) -> bool | None:  # noqa: FBT001
     return check_and_get_environ_var(
         f"ANNDATA_{option.upper()}",
-        str(int(default_value)),
-        None,
-        lambda x: int(x),
+        "" if default_value is None else str(int(default_value)),
+        ["0", "1", ""],
+        lambda x: None if x == "" else bool(int(x)),
+    )
+
+
+def check_and_get_int(option: str, default_value: int) -> int:
+    return check_and_get_environ_var(
+        f"ANNDATA_{option.upper()}", str(int(default_value)), None, int
     )
 
 
@@ -200,13 +204,13 @@ class SettingsManager:
         )
 
     @old_positionals("default_value", "description", "validate", "option_type")
-    def register(
+    def register[T](
         self,
         option: str,
         *,
         default_value: T,
         description: str,
-        validate: Callable[[T], None],
+        validate: Callable[[T, Self], None],
         option_type: object | None = None,
         get_from_env: Callable[[str, T], T] = lambda x, y: y,
     ) -> None:
@@ -229,7 +233,7 @@ class SettingsManager:
             Default behavior is to return `default_value` without checking the environment.
         """
         try:
-            validate(default_value)
+            validate(default_value, self)
         except (ValueError, TypeError) as e:
             e.add_note(f"for option {option!r}")
             raise e
@@ -307,7 +311,7 @@ class SettingsManager:
             )
             raise AttributeError(msg)
         registered_option = self._registered_options[option]
-        registered_option.validate(val)
+        registered_option.validate(val, self)
         self._config[option] = val
 
     def __getattr__(self, option: str) -> object:
@@ -326,7 +330,7 @@ class SettingsManager:
         if option in self._deprecated_options:
             deprecated = self._deprecated_options[option]
             msg = f"{option!r} will be removed in {deprecated.removal_version}. {deprecated.message}"
-            warnings.warn(msg, FutureWarning, stacklevel=2)
+            warn(msg, FutureWarning)
         if option in self._config:
             return self._config[option]
         msg = f"{option} not found."
@@ -364,10 +368,13 @@ class SettingsManager:
         """
         restore = {a: getattr(self, a) for a in overrides}
         try:
-            for attr, value in overrides.items():
-                setattr(self, attr, value)
+            # Preserve order so that settings that depend on each other can be overridden together i.e., always override zarr version before sharding
+            for k in self._config:
+                if k in overrides:
+                    setattr(self, k, overrides.get(k))
             yield None
         finally:
+            # TODO: does the order need to be preserved when restoring?
             for attr, value in restore.items():
                 setattr(self, attr, value)
 
@@ -391,11 +398,12 @@ settings = SettingsManager()
 ##################################################################################
 # PLACE REGISTERED SETTINGS HERE SO THEY CAN BE PICKED UP FOR DOCSTRING CREATION #
 ##################################################################################
-V = TypeVar("V")
 
 
-def gen_validator(_type: type[V]) -> Callable[[V], None]:
-    def validate_type(val: V) -> None:
+def gen_validator[V](
+    _type: type[V] | tuple[type[V], ...], /
+) -> Callable[[V, SettingsManager], None]:
+    def validate_type(val: V, settings: SettingsManager) -> None:
         if not isinstance(val, _type):
             msg = f"{val} not valid {_type}"
             raise TypeError(msg)
@@ -427,20 +435,32 @@ settings.register(
 
 settings.register(
     "allow_write_nullable_strings",
-    default_value=False,
-    description="Whether or not to allow writing of `pd.arrays.StringArray`.",
-    validate=validate_bool,
-    get_from_env=check_and_get_bool,
+    default_value=None,
+    description=(
+        "Whether or not to allow writing of `pd.arrays.[Arrow]StringArray`. "
+        "When set to `None`, it will be inferred from `pd.options.future.infer_string`. "
+        "When set to `False` explicitly, we will try writing `string` arrays in the old, non-nullable format."
+    ),
+    validate=gen_validator((bool, NoneType)),
+    option_type=bool | None,
+    get_from_env=check_and_get_bool_or_none,
 )
 
 
-def validate_zarr_write_format(format: int):
-    validate_int(format)
+def validate_zarr_write_format(format: int, settings: SettingsManager):
+    validate_int(format, settings)
     if format not in {2, 3}:
         msg = "non-v2 zarr on-disk format not supported"
         raise ValueError(msg)
-    if format == 3 and is_zarr_v2():
-        msg = "Cannot write v3 format against v2 package"
+    if format == 2 and getattr(settings, "auto_shard_zarr_v3", False):
+        msg = "Cannot set `zarr_write_format` to 2 with autosharding on.  Please set to `False` `anndata.settings.auto_shard_zarr_v3`"
+        raise ValueError(msg)
+
+
+def validate_zarr_sharding(auto_shard: bool, settings: SettingsManager):  # noqa: FBT001
+    validate_bool(auto_shard, settings)
+    if auto_shard and settings.zarr_write_format == 2:
+        msg = "Cannot shard v2 format data. Please set `anndata.settings.zarr_write_format` to 3."
         raise ValueError(msg)
 
 
@@ -450,16 +470,13 @@ settings.register(
     description="Which version of zarr to write to when anndata must internally open a write-able zarr group.",
     validate=validate_zarr_write_format,
     get_from_env=lambda name, default: check_and_get_environ_var(
-        f"ANNDATA_{name.upper()}",
-        str(default),
-        ["2", "3"],
-        lambda x: int(x),
+        f"ANNDATA_{name.upper()}", str(default), ["2", "3"], int
     ),
 )
 
 
-def validate_sparse_settings(val: Any) -> None:
-    validate_bool(val)
+def validate_sparse_settings(val: Any, settings: SettingsManager) -> None:
+    validate_bool(val, settings)
 
 
 settings.register(
@@ -476,6 +493,43 @@ settings.register(
     description="Minimum number of rows at a time to copy when writing out an H5 Dataset to a new location",
     validate=validate_int,
     get_from_env=check_and_get_int,
+)
+
+settings.register(
+    "disallow_forward_slash_in_h5ad",
+    default_value=False,
+    description="Whether or not to disallow the `/` character in keys for h5ad files",
+    validate=validate_bool,
+    get_from_env=check_and_get_bool,
+)
+
+settings.register(
+    "write_csr_csc_indices_with_min_possible_dtype",
+    default_value=False,
+    description="Write a csr or csc matrix with the minimum possible data type for `indices`, always unsigned integer.",
+    validate=validate_bool,
+    get_from_env=check_and_get_bool,
+)
+
+settings.register(
+    "auto_shard_zarr_v3",
+    default_value=False,
+    description="Whether or not to use zarr's auto computation of sharding for v3.  For v2 this setting will be ignored. The setting will apply to all calls to anndata's writing mechanism (write_zarr / write_elem) and will **not** override any user-defined kwargs for shards.",
+    validate=validate_zarr_sharding,
+    get_from_env=check_and_get_bool,
+)
+
+
+settings.register(
+    "copy_on_write_X",
+    default_value=False,
+    description=(
+        "Whether to copy-on-write X. "
+        "Currently `my_adata_view[subset].X = value` will write back to the original AnnData object at the `subset` location. "
+        "`X` is the only element where this behavior is implemented though."
+    ),
+    validate=validate_bool,
+    get_from_env=check_and_get_bool,
 )
 
 
