@@ -1,9 +1,11 @@
 from __future__ import annotations
 
 import re
+import sys
 import warnings
-from functools import singledispatch, wraps
-from typing import TYPE_CHECKING
+from functools import partial, singledispatch
+from types import FunctionType, UnionType
+from typing import TYPE_CHECKING, Literal, TypeAliasType, get_args, get_origin
 
 import h5py
 import numpy as np
@@ -13,12 +15,13 @@ from scipy import sparse
 import anndata
 
 from ._core.sparse_dataset import BaseCompressedSparseDataset
+from ._warnings import warn
 from .compat import CSArray, CupyArray, CupySparseMatrix, DaskArray
 from .logging import get_logger
 
 if TYPE_CHECKING:
-    from collections.abc import Iterable, Mapping, Sequence
-    from typing import Any, Literal
+    from collections.abc import Callable, Iterable, Mapping, Sequence
+    from typing import Any, LiteralString
 
 logger = get_logger(__name__)
 
@@ -216,7 +219,7 @@ except ImportError:
     pass
 
 
-def make_index_unique(index: pd.Index, join: str = "-"):
+def make_index_unique(index: pd.Index[str], join: str = "-") -> pd.Index[str]:
     """
     Makes the index unique by appending a number string to each duplicate index element:
     '1', '2', etc.
@@ -235,18 +238,18 @@ def make_index_unique(index: pd.Index, join: str = "-"):
     --------
     >>> from anndata import AnnData
     >>> adata = AnnData(np.ones((2, 3)), var=pd.DataFrame(index=["a", "a", "b"]))
-    >>> adata.var_names
-    Index(['a', 'a', 'b'], dtype='object')
+    >>> adata.var_names.astype("string")
+    Index(['a', 'a', 'b'], dtype='string')
     >>> adata.var_names_make_unique()
-    >>> adata.var_names
-    Index(['a', 'a-1', 'b'], dtype='object')
+    >>> adata.var_names.astype("string")
+    Index(['a', 'a-1', 'b'], dtype='string')
     """
     if index.is_unique:
         return index
     from collections import Counter
 
-    values = index.values.copy()
-    indices_dup = index.duplicated(keep="first")
+    values = index.array.copy()
+    indices_dup = index.duplicated(keep="first") & ~index.isna()
     values_dup = values[indices_dup]
     values_set = set(values)
     counter = Counter()
@@ -272,8 +275,7 @@ def make_index_unique(index: pd.Index, join: str = "-"):
             "Example key collisions generated by the make_index_unique algorithm: "
             f"{example_colliding_values}"
         )
-        # 3: caller -> 2: `{obs,var}_names_make_unique` -> 1: here
-        warnings.warn(msg, UserWarning, stacklevel=3)
+        warn(msg, UserWarning)
     values[indices_dup] = values_dup
     index = pd.Index(values, name=index.name)
     return index
@@ -292,12 +294,11 @@ def join_english(words: Iterable[str], conjunction: str = "or") -> str:
 
 def warn_names_duplicates(attr: str):
     names = "Observation" if attr == "obs" else "Variable"
-    warnings.warn(
+    msg = (
         f"{names} names are not unique. "
-        f"To make them unique, call `.{attr}_names_make_unique`.",
-        UserWarning,
-        stacklevel=2,
+        f"To make them unique, call `.{attr}_names_make_unique`."
     )
+    warn(msg, UserWarning)
 
 
 def ensure_df_homogeneous(
@@ -310,8 +311,7 @@ def ensure_df_homogeneous(
         arr = df.to_numpy()
     if df.dtypes.nunique() != 1:
         msg = f"{name} converted to numpy array with dtype {arr.dtype}"
-        # 4: caller -> 3: `AnnData.__init__` -> 2: `_init_as_actual` → 1: here
-        warnings.warn(msg, UserWarning, stacklevel=4)
+        warn(msg, UserWarning)
     return arr
 
 
@@ -358,44 +358,33 @@ def convert_dictionary_to_structured_array(source: Mapping[str, Sequence[Any]]):
     return arr
 
 
-def warn_once(msg: str, category: type[Warning], stacklevel: int = 1):
-    warnings.warn(msg, category, stacklevel=stacklevel)
+def warn_once(msg: str, category: type[Warning]) -> None:
+    warn(msg, category)
     # Prevent from showing up every time an awkward array is used
     # You'd think `'once'` works, but it doesn't at the repl and in notebooks
     warnings.filterwarnings("ignore", category=category, message=re.escape(msg))
 
 
-def deprecated(
-    new_name: str,
-    category: type[Warning] = FutureWarning,
-    add_msg: str = "",
-    *,
-    hide: bool = True,
-):
-    """\
-    This is a decorator which can be used to mark functions
-    as deprecated with a FutureWarning. It will result in a warning being emitted
-    when the function is used.
-    """
+if TYPE_CHECKING:
+    from warnings import deprecated
+else:
+    if sys.version_info >= (3, 13):
+        from warnings import deprecated as _deprecated
+    else:
+        from typing_extensions import deprecated as _deprecated
+    deprecated = partial(_deprecated, category=FutureWarning)
 
-    def decorator(func):
-        name = func.__qualname__
-        msg = (
-            f"Use {new_name} instead of {name}, "
-            f"{name} is deprecated and will be removed in the future."
-        )
-        if add_msg:
-            msg += f" {add_msg}"
 
-        @wraps(func)
-        def new_func(*args, **kwargs):
-            warnings.warn(msg, category=category, stacklevel=2)
-            return func(*args, **kwargs)
-
-        setattr(new_func, "__deprecated", (category, msg, hide))
-        return new_func
-
-    return decorator
+def deprecation_msg(
+    name: LiteralString, new_name: LiteralString, add_msg: LiteralString | None = None
+) -> LiteralString:
+    msg = (
+        f"Use {new_name} instead of {name}, "
+        f"{name} is deprecated and will be removed in the future."
+    )
+    if add_msg is not None:
+        msg += f" {add_msg}"
+    return msg
 
 
 class DeprecationMixinMeta(type):
@@ -405,17 +394,47 @@ class DeprecationMixinMeta(type):
     """
 
     def __dir__(cls):
-        def is_hidden(attr) -> bool:
+        dont_hide = getattr(cls, "_DONT_HIDE_DEPRECATED", set())
+
+        def is_hidden(attr: object) -> bool:
             if isinstance(attr, property):
                 attr = attr.fget
-            _, _, hide = getattr(attr, "__deprecated", (None, None, False))
-            return hide
+            is_deprecated = bool(getattr(attr, "__deprecated__", None))
+            return is_deprecated and getattr(attr, "__name__", None) not in dont_hide
 
         return [
             item
             for item in type.__dir__(cls)
             if not is_hidden(getattr(cls, item, None))
         ]
+
+
+def set_module[C: FunctionType | type](name: str, /) -> Callable[[C], C]:
+    def decorator(f: C) -> C:
+        f.__module__ = name
+        return f
+
+    return decorator
+
+
+def get_union_members[T](
+    typ: TypeAliasType | UnionType | T,
+    can_get_args: Callable[[object], bool] = lambda _: False,
+) -> tuple[T, ...]:
+    """Get args of a nested union of types."""
+    while isinstance(typ, TypeAliasType):
+        typ = typ.__value__
+    if not isinstance(typ, UnionType) and not can_get_args(typ):
+        return (typ,)
+    return tuple(
+        t for sub in get_args(typ) for t in get_union_members(sub, can_get_args)
+    )
+
+
+get_literal_members = partial(
+    get_union_members, can_get_args=lambda x: get_origin(x) is Literal
+)
+"""Get args of a nested union of Literal types."""
 
 
 def raise_value_error_if_multiindex_columns(df: pd.DataFrame, attr: str):
@@ -440,7 +459,7 @@ def module_get_attr_redirect(
             f"Importing {attr_name} from `{full_old_module_path}` is deprecated. "
             f"Import anndata.{new_path} instead."
         )
-        warnings.warn(msg, FutureWarning, stacklevel=2)
+        warn(msg, FutureWarning)
         # hacky import_object_by_name, but we test all these
         mod = anndata
         while "." in new_path:
