@@ -17,7 +17,7 @@ if TYPE_CHECKING:
     from os import PathLike
     from typing import Literal
 
-    from .._types import ArrayStorageType
+    from .._types import _ArrayStorageType
     from . import anndata
 
 
@@ -27,15 +27,24 @@ class AnnDataFileManager:
     def __init__(
         self,
         adata: anndata.AnnData,
-        filename: PathLike[str] | str | None = None,
-        filemode: Literal["r", "r+"] | None = None,
+        file_name: PathLike[str] | str | None = None,
+        file_mode: Literal["r", "r+"] | None = None,
+        file_obj: h5py.File | None = None,
     ):
+        if file_obj is not None and (file_name is not None or file_mode is not None):
+            msg = "Cannot provide both a h5py.File and the name and/or mode arguments to constructor"
+            raise ValueError(msg)
         self._adata_ref = weakref.ref(adata)
-        self.filename = filename
-        self._filemode = filemode
-        self._file = None
-        if filename:
-            self.open()
+        if file_obj is not None:
+            self.filename = filename(file_obj)
+            self._filemode = file_obj.mode
+            self._file = file_obj
+        else:
+            self.filename = file_name
+            self._filemode = file_mode
+            self._file = file_obj
+            if file_name and not self._file:
+                self.open()
 
     def __getstate__(self):
         state = self.__dict__.copy()
@@ -82,16 +91,16 @@ class AnnDataFileManager:
         return self._filename
 
     @filename.setter
-    def filename(self, filename: PathLike[str] | str | None):
-        self._filename = None if filename is None else Path(filename)
+    def filename(self, file_name: PathLike[str] | str | None):
+        self._filename = None if file_name is None else Path(file_name)
 
     def open(
         self,
-        filename: PathLike[str] | str | None = None,
+        file_name: PathLike[str] | str | None = None,
         filemode: Literal["r", "r+"] | None = None,
     ):
-        if filename is not None:
-            self.filename = filename
+        if file_name is not None:
+            self.filename = file_name
         if filemode is not None:
             self._filemode = filemode
         if self.filename is None:
@@ -134,7 +143,7 @@ def to_memory(x, *, copy: bool = False):
 
 @to_memory.register(ZarrArray)
 @to_memory.register(h5py.Dataset)
-def _(x: ArrayStorageType, *, copy: bool = False):
+def _(x: _ArrayStorageType, *, copy: bool = False):
     return x[...]
 
 
