@@ -4,19 +4,17 @@ Code for merging/ concatenating AnnData objects.
 
 from __future__ import annotations
 
+import uuid
 from collections import OrderedDict
 from collections.abc import Callable, Mapping, MutableSet
 from functools import partial, reduce, singledispatch
 from itertools import repeat
 from operator import and_, or_, sub
-from typing import TYPE_CHECKING, Literal, TypeVar
-from warnings import warn
+from typing import TYPE_CHECKING, Literal, cast
 
 import numpy as np
 import pandas as pd
-import scipy
 from natsort import natsorted
-from packaging.version import Version
 from scipy import sparse
 
 from anndata._core.file_backing import to_memory
@@ -30,9 +28,9 @@ from ..compat import (
     CupyCSRMatrix,
     CupySparseMatrix,
     DaskArray,
-    _map_cat_to_str,
+    has_xp,
 )
-from ..utils import asarray, axis_len, warn_once
+from ..utils import asarray, axis_len, warn, warn_once
 from .anndata import AnnData
 from .index import _subset, make_slice
 from .xarray import Dataset2D
@@ -41,13 +39,14 @@ if TYPE_CHECKING:
     from collections.abc import Collection, Generator, Iterable, Sequence
     from typing import Any
 
+    from numpy.typing import NDArray
     from pandas.api.extensions import ExtensionDtype
 
     from anndata._types import Join_T
 
-    from ..compat import XDataArray, XDataset
+    from ..compat import XDataArray
+    from ..types import SupportsArrayApi
 
-T = TypeVar("T")
 
 ###################
 # Utilities
@@ -146,11 +145,16 @@ def equal_dask_array(a, b) -> bool:
         return False
     if isinstance(b, DaskArray) and tokenize(a) == tokenize(b):
         return True
-    if isinstance(a._meta, CSMatrix):
+    if isinstance(a._meta, np.ndarray):
+        return da.equal(a, b, where=~(da.isnan(a) & da.isnan(b))).all().compute()
+    if a.chunksize == b.chunksize and isinstance(
+        a._meta, CupySparseMatrix | CSMatrix | CSArray
+    ):
         # TODO: Maybe also do this in the other case?
         return da.map_blocks(equal, a, b, drop_axis=(0, 1)).all()
-    else:
-        return da.equal(a, b, where=~(da.isnan(a) == da.isnan(b))).all()
+    msg = "Misaligned chunks detected when checking for merge equality of dask arrays.  Reading full arrays into memory."
+    warn(msg, UserWarning)
+    return equal(a.compute(), b.compute())
 
 
 @equal.register(np.ndarray)
@@ -185,15 +189,6 @@ def equal_sparse(a, b) -> bool:
             # Comparison broken for CSC matrices
             # https://github.com/cupy/cupy/issues/7757
             a, b = CupyCSRMatrix(a), CupyCSRMatrix(b)
-        if Version(scipy.__version__) >= Version("1.16.0rc1"):
-            # TODO: https://github.com/scipy/scipy/issues/23068
-            return bool(
-                a.format == b.format
-                and (a.shape == b.shape)
-                and np.all(a.indptr == b.indptr)
-                and np.all(a.indices == b.indices)
-                and np.all((a.data == b.data) | (np.isnan(a.data) & np.isnan(b.data)))
-            )
         comp = a != b
         if isinstance(comp, bool):
             return not comp
@@ -249,106 +244,89 @@ def as_cp_sparse(x) -> CupySparseMatrix:
 def unify_dtypes(
     dfs: Iterable[pd.DataFrame | Dataset2D],
 ) -> list[pd.DataFrame | Dataset2D]:
-    """
-    Attempts to unify datatypes from multiple dataframes.
+    """Attempt to unify datatypes from multiple dataframes.
 
     For catching cases where pandas would convert to object dtype.
     """
     dfs = list(dfs)
     # Get shared categorical columns
-    df_dtypes = [dict(df.dtypes) for df in dfs]
+    df_dtypes = [
+        cast("pd.Series[ExtensionDtype]", df.dtypes).to_dict()
+        if isinstance(df, pd.DataFrame)
+        else df.dtypes
+        for df in dfs
+    ]
     columns = reduce(lambda x, y: x.union(y), [df.columns for df in dfs])
-
-    dtypes: dict[str, list[np.dtype | ExtensionDtype]] = {col: [] for col in columns}
-    for col in columns:
-        for df in df_dtypes:
-            dtypes[col].append(df.get(col, None))
-
+    dtypes = {
+        col: (
+            [df[col] for df in df_dtypes if col in df],
+            any(col not in df for df in df_dtypes),
+        )
+        for col in columns
+    }
     if len(dtypes) == 0:
         return dfs
-    else:
-        dfs = [df.copy(deep=False) for df in dfs]
 
     new_dtypes = {
         col: target_dtype
-        for col, dtype in dtypes.items()
-        if (target_dtype := try_unifying_dtype(dtype)) is not None
+        for col, (dts, has_missing) in dtypes.items()
+        if (target_dtype := try_unifying_dtype(dts, has_missing=has_missing))
+        is not None
     }
 
+    dfs = [df.copy(deep=False) for df in dfs]
     for df in dfs:
         for col, dtype in new_dtypes.items():
             if col in df:
                 df[col] = df[col].astype(dtype)
-
     return dfs
 
 
-def try_unifying_dtype(  # noqa PLR0911, PLR0912
-    col: Sequence[np.dtype | ExtensionDtype],
-) -> pd.core.dtypes.base.ExtensionDtype | None:
-    """
-    If dtypes can be unified, returns the dtype they would be unified to.
+def try_unifying_dtype(
+    dtypes: Sequence[np.dtype | ExtensionDtype], *, has_missing: bool
+) -> ExtensionDtype | type[object] | None:
+    """Determine unified dtype if possible.
 
-    Returns None if they can't be unified, or if we can expect pandas to unify them for
-    us.
+    Returns None if they can’t be unified, or if we can expect pandas to unify them for us.
 
     Params
     ------
-    col:
-        A list of dtypes to unify. Can be numpy/ pandas dtypes, or None (which denotes
-        a missing value)
+    dtypes
+        A list of dtypes to unify. Can be numpy or pandas dtypes
+    has_missing
+        Whether the result needs to accommodate missing values
     """
-    dtypes: set[pd.CategoricalDtype] = set()
     # Categorical
-    if any(isinstance(dtype, pd.CategoricalDtype) for dtype in col):
-        ordered = False
-        for dtype in col:
-            if isinstance(dtype, pd.CategoricalDtype):
-                dtypes.add(dtype)
-                ordered = ordered | dtype.ordered
-            elif not pd.isnull(dtype):
-                return None
-        if len(dtypes) > 0:
-            categories = reduce(
-                lambda x, y: x.union(y),
-                (dtype.categories for dtype in dtypes if not pd.isnull(dtype)),
-            )
-
-            if not ordered:
-                return pd.CategoricalDtype(natsorted(categories), ordered=False)
-            else:  # for xarray Datasets, see https://github.com/pydata/xarray/issues/10247
-                categories_intersection = reduce(
-                    lambda x, y: x.intersection(y),
-                    (
-                        dtype.categories
-                        for dtype in dtypes
-                        if not pd.isnull(dtype) and len(dtype.categories) > 0
-                    ),
-                )
-                if len(categories_intersection) < len(categories):
-                    return object
-                else:
-                    same_orders = all(
-                        dtype.ordered
-                        for dtype in dtypes
-                        if not pd.isnull(dtype) and len(dtype.categories) > 0
-                    )
-                    same_orders &= all(
-                        np.all(categories == dtype.categories)
-                        for dtype in dtypes
-                        if not pd.isnull(dtype) and len(dtype.categories) > 0
-                    )
-                    if same_orders:
-                        return next(iter(dtypes))
-                    return object
-    # Boolean
-    elif all(pd.api.types.is_bool_dtype(dtype) or dtype is None for dtype in col):
-        if any(dtype is None for dtype in col):
-            return pd.BooleanDtype()
-        else:
+    if any(isinstance(dtype, pd.CategoricalDtype) for dtype in dtypes):
+        if not all(isinstance(dtype, pd.CategoricalDtype) for dtype in dtypes):
             return None
-    else:
-        return None
+        if TYPE_CHECKING:
+            dtypes = cast("Sequence[pd.CategoricalDtype]", dtypes)
+
+        all_categories = reduce(
+            lambda x, y: x.union(y), (dtype.categories for dtype in dtypes)
+        )
+        if not any(dtype.ordered for dtype in dtypes):
+            return pd.CategoricalDtype(natsorted(all_categories), ordered=False)
+
+        dtypes_with_categories = [
+            dtype for dtype in dtypes if len(dtype.categories) > 0
+        ]
+        if dtypes_with_categories and all(
+            len(dtype.categories) == len(all_categories)
+            and dtype.ordered
+            and np.all(all_categories == dtype.categories)
+            for dtype in dtypes_with_categories
+        ):
+            return dtypes_with_categories[0]
+
+        return object
+
+    # Boolean
+    if all(pd.api.types.is_bool_dtype(dtype) for dtype in dtypes) and has_missing:
+        return pd.BooleanDtype()
+
+    return None
 
 
 def check_combinable_cols(cols: list[pd.Index], join: Join_T):
@@ -435,7 +413,7 @@ def _dask_block_diag(mats):
 ###################
 
 
-def unique_value(vals: Collection[T]) -> T | MissingVal:
+def unique_value[T](vals: Collection[T]) -> T | MissingVal:
     """
     Given a collection vals, returns the unique value (if one exists), otherwise
     returns MissingValue.
@@ -447,9 +425,9 @@ def unique_value(vals: Collection[T]) -> T | MissingVal:
     return unique_val
 
 
-def first(vals: Collection[T]) -> T | MissingVal:
+def first[T](vals: Collection[T]) -> T | MissingVal:
     """
-    Given a collection of vals, return the first non-missing one.If they're all missing,
+    Given a collection of vals, return the first non-missing one. If they're all missing,
     return MissingVal.
     """
     for val in vals:
@@ -458,7 +436,7 @@ def first(vals: Collection[T]) -> T | MissingVal:
     return MissingVal
 
 
-def only(vals: Collection[T]) -> T | MissingVal:
+def only[T](vals: Collection[T]) -> T | MissingVal:
     """Return the only value in the collection, otherwise MissingVal."""
     if len(vals) == 1:
         return vals[0]
@@ -516,9 +494,6 @@ def merge_only(ds: Collection[Mapping]) -> Mapping:
 # Interface
 ###################
 
-# Leaving out for now, it's ugly in the rendered docs and would be adding a dependency.
-# from typing_extensions import Literal
-# UNS_STRATEGIES_TYPE = Literal[None, "same", "unique", "first", "only"]
 MERGE_STRATEGIES = {
     None: lambda x: {},
     "same": merge_same,
@@ -560,7 +535,7 @@ class Reindexer:
         Together with `old_pos` this forms a mapping.
     """
 
-    def __init__(self, old_idx, new_idx):
+    def __init__(self, old_idx: pd.Index, new_idx: pd.Index) -> None:
         self.old_idx = old_idx
         self.new_idx = new_idx
         self.no_change = new_idx.equals(old_idx)
@@ -594,8 +569,11 @@ class Reindexer:
             return self._apply_to_dask_array(el, axis=axis, fill_value=fill_value)
         elif isinstance(el, CupyArray):
             return self._apply_to_cupy_array(el, axis=axis, fill_value=fill_value)
-        else:
-            return self._apply_to_array(el, axis=axis, fill_value=fill_value)
+        elif has_xp(el):
+            return self._apply_to_array_api(el, axis=axis, fill_value=fill_value)
+        else:  # pragma: no cover
+            msg = "Cannot reindex element of unsupported type."
+            raise TypeError(msg)
 
     def _apply_to_df_like(self, el: pd.DataFrame | Dataset2D, *, axis, fill_value=None):
         if fill_value is None:
@@ -617,6 +595,9 @@ class Reindexer:
         sub_el = _subset(el, make_slice(indexer, axis, len(shape)))
 
         if any(indexer == -1):
+            # TODO: Remove this condition once https://github.com/dask/dask/pull/12078 is released
+            if isinstance(sub_el._meta, CSArray | CSMatrix) and np.isscalar(fill_value):
+                fill_value = np.array([[fill_value]])
             sub_el[make_slice(indexer == -1, axis, len(shape))] = fill_value
 
         return sub_el
@@ -647,21 +628,34 @@ class Reindexer:
 
         return out
 
-    def _apply_to_array(self, el, *, axis, fill_value=None):
+    def _apply_to_array_api(
+        self, el: SupportsArrayApi, *, axis: int, fill_value=None
+    ) -> SupportsArrayApi:
         if fill_value is None:
             fill_value = default_fill_value([el])
+        xp = el.__array_namespace__()
+        indexer = xp.asarray(self.idx)
+
+        # Handling edge case to mimic pandas behavior
         if el.shape[axis] == 0:
-            # Presumably faster since it won't allocate the full array
             shape = list(el.shape)
-            shape[axis] = len(self.new_idx)
-            return np.broadcast_to(fill_value, tuple(shape))
-
-        indexer = self.idx
-
-        # Indexes real fast, and does outer indexing
-        return pd.api.extensions.take(
-            el, indexer, axis=axis, allow_fill=True, fill_value=fill_value
-        )
+            shape[axis] = indexer.shape[0]
+            # convert fill_value to the same type as el - to keep everything in the same dtype
+            # fv = xp.asarray(fill_value, dtype=getattr(el, "dtype", None))
+            fv = xp.asarray(fill_value)
+            return xp.broadcast_to(fv, shape)
+        # marking which positions are missing, so we could use fill_value
+        missing_mask = indexer == -1
+        safe_indexer = xp.where(missing_mask, 0, indexer)
+        taken = xp.take(el, safe_indexer, axis=axis)
+        # Expand mask so we can apply xp.where along the right axis
+        shape = [1] * taken.ndim
+        shape[axis] = missing_mask.shape[0]
+        mask = missing_mask.reshape(shape)
+        if not xp.any(missing_mask):
+            return taken
+        fv = xp.asarray(fill_value)
+        return xp.where(mask, fv, taken)
 
     def _apply_to_sparse(  # noqa: PLR0912
         self, el: CSMatrix | CSArray, *, axis, fill_value=None
@@ -757,7 +751,7 @@ class Reindexer:
             return el[self.idx]
 
     @property
-    def idx(self):
+    def idx(self) -> NDArray[np.intp]:
         return self.old_idx.get_indexer(self.new_idx)
 
 
@@ -786,7 +780,7 @@ def default_fill_value(els):
         return np.nan
 
 
-def gen_reindexer(new_var: pd.Index, cur_var: pd.Index):
+def gen_reindexer(new_var: pd.Index, cur_var: pd.Index) -> Reindexer:
     """
     Given a new set of var_names, and a current set, generates a function which will reindex
     a matrix to be aligned with the new set.
@@ -905,6 +899,22 @@ def concat_arrays(  # noqa: PLR0911, PLR0912
             format="csr",
         )
         return mat
+
+    elif all(has_xp(a) for a in arrays):
+        # All arrays are array-api compatible
+
+        # use first as a reference to check if all of the arrays are the same type
+        if len(array_apis := {a.__array_namespace__() for a in arrays}) > 1:
+            msg = f"Cannot concatenate array-api arrays from different backends: {array_apis}."
+            raise ValueError(msg)
+
+        return next(iter(array_apis)).concatenate(
+            [
+                f(x, fill_value=fill_value, axis=1 - axis)
+                for f, x in zip(reindexers, arrays, strict=True)
+            ],
+            axis=axis,
+        )
     else:
         return np.concatenate(
             [
@@ -943,7 +953,7 @@ def inner_concat_aligned_mapping(
     return result
 
 
-def gen_inner_reindexers(els, new_index, axis: Literal[0, 1] = 0):
+def gen_inner_reindexers(els, new_index, axis: Literal[0, 1] = 0) -> list[Reindexer]:
     alt_axis = 1 - axis
     if axis == 0:
         df_indices = lambda x: x.columns
@@ -960,8 +970,13 @@ def gen_inner_reindexers(els, new_index, axis: Literal[0, 1] = 0):
             msg = "Cannot concatenate an AwkwardArray with other array types."
             raise NotImplementedError(msg)
         common_keys = intersect_keys(el.fields for el in els)
+        # TODO: replace dtype=object once this is fixed: https://github.com/scikit-hep/awkward/issues/3730
         reindexers = [
-            Reindexer(pd.Index(el.fields), pd.Index(list(common_keys))) for el in els
+            Reindexer(
+                pd.Index(el.fields, dtype=object),
+                pd.Index(list(common_keys), dtype=object),
+            )
+            for el in els
         ]
     else:
         min_ind = min(el.shape[alt_axis] for el in els)
@@ -986,12 +1001,12 @@ def gen_outer_reindexers(els, shapes, new_index: pd.Index, *, axis=0):
         if not all(isinstance(el, AwkArray) for el in els if not_missing(el)):
             msg = "Cannot concatenate an AwkwardArray with other array types."
             raise NotImplementedError(msg)
-        warn_once(
+        msg = (
             "Outer joins on awkward.Arrays will have different return values in the future. "
             "For details, and to offer input, please see:\n\n\t"
-            "https://github.com/scverse/anndata/issues/898",
-            ExperimentalFeatureWarning,
+            "https://github.com/scverse/anndata/issues/898"
         )
+        warn_once(msg, ExperimentalFeatureWarning)
         # all_keys = union_keys(el.fields for el in els if not_missing(el))
         reindexers = []
         for el in els:
@@ -1020,7 +1035,7 @@ def missing_element(
     axis: Literal[0, 1] = 0,
     fill_value: Any | None = None,
     off_axis_size: int = 0,
-) -> np.ndarray | DaskArray:
+) -> NDArray[np.bool_] | DaskArray:
     """Generates value to use when there is a missing element."""
     should_return_dask = any(isinstance(el, DaskArray) for el in els)
     # 0 sized array for in-memory prevents allocating unnecessary memory while preserving broadcasting.
@@ -1031,7 +1046,16 @@ def missing_element(
         return da.full(
             shape, default_fill_value(els) if fill_value is None else fill_value
         )
-    return np.zeros(shape, dtype=bool)
+    non_numpy_array_apis = {
+        a.__array_namespace__()
+        for a in els
+        if has_xp(a) and not isinstance(a, np.ndarray)
+    }
+    if len(non_numpy_array_apis) not in {0, 1}:
+        msg = "Cannot generate missing elements when there are multiple array backends supported including at least one that is array api compatible."
+        raise ValueError(msg)
+    xp = next(iter(non_numpy_array_apis)) if len(non_numpy_array_apis) == 1 else np
+    return xp.zeros(shape, dtype=bool)
 
 
 def outer_concat_aligned_mapping(
@@ -1182,7 +1206,6 @@ def make_dask_col_from_extension_dtype(
         get_chunksize,
         maybe_open_h5,
     )
-    from anndata.compat import XDataArray
     from anndata.compat import xarray as xr
     from anndata.experimental import read_elem_lazy
 
@@ -1192,7 +1215,6 @@ def make_dask_col_from_extension_dtype(
         base_path_or_zarr_group is not None and elem_name is not None
     ):  # lazy, backed by store
         dims = col.dims
-        coords = col.coords.copy()
         with maybe_open_h5(base_path_or_zarr_group, elem_name) as f:
             maybe_chunk_size = get_chunksize(read_elem_lazy(f))
             chunk_size = (
@@ -1209,16 +1231,11 @@ def make_dask_col_from_extension_dtype(
                 variable = xr.Variable(
                     data=xr.core.indexing.LazilyIndexedArray(v), dims=dims
                 )
-                data_array = XDataArray(
-                    variable,
-                    coords=coords,
-                    dims=dims,
-                )
                 idx = tuple(
                     slice(start, stop)
                     for start, stop in block_info[None]["array-location"]
                 )
-                chunk = np.array(data_array.data[idx])
+                chunk = np.array(variable.data[idx])
             return chunk
 
         if col.dtype == "category" or col.dtype == "string" or use_only_object_dtype:  # noqa PLR1714
@@ -1230,6 +1247,7 @@ def make_dask_col_from_extension_dtype(
             chunks=chunk_size,
             meta=np.array([], dtype=dtype),
             dtype=dtype,
+            name=f"{uuid.uuid4()}/{base_path_or_zarr_group}/{elem_name}-{dtype}",
         )
 
     return da.from_array(col.values, chunks=-1)  # in-memory
@@ -1237,7 +1255,7 @@ def make_dask_col_from_extension_dtype(
 
 def make_xarray_extension_dtypes_dask(
     annotations: Iterable[Dataset2D], *, use_only_object_dtype: bool = False
-) -> Generator[XDataset, None, None]:
+) -> Generator[Dataset2D, None, None]:
     """
     Creates a generator of Dataset2D objects with dask arrays in place of :class:`pandas.api.extensions.ExtensionArray` dtype columns.
 
@@ -1300,9 +1318,10 @@ def concat_dataset2d_on_annot_axis(
     -------
     Concatenated :class:`~anndata.experimental.backed.Dataset2D`
     """
+    import xarray as xr
+
     from anndata._core.xarray import Dataset2D
     from anndata._io.specs.lazy_methods import DUMMY_RANGE_INDEX_KEY
-    from anndata.compat import xarray as xr
 
     annotations_re_indexed = []
     have_backed = any(a.is_backed for a in annotations)
@@ -1503,16 +1522,19 @@ def concat(  # noqa: PLR0912, PLR0913, PLR0915
     AnnData object with n_obs × n_vars = 4 × 2
         obs: 'group'
         layers: None
-    >>> (inner.obs_names, inner.var_names)  # doctest: +NORMALIZE_WHITESPACE
-    (Index(['s1', 's2', 's3', 's4'], dtype='object'),
-    Index(['var1', 'var2'], dtype='object'))
+    >>> (
+    ...     inner.obs_names.astype("string"),
+    ...     inner.var_names.astype("string"),
+    ... )  # doctest: +NORMALIZE_WHITESPACE
+    (Index(['s1', 's2', 's3', 's4'], dtype='string'),
+    Index(['var1', 'var2'], dtype='string'))
     >>> outer = ad.concat([a, b], join="outer")  # Joining on union of variables
     >>> outer
     AnnData object with n_obs × n_vars = 4 × 3
         obs: 'group', 'measure'
-        layers: None
-    >>> outer.var_names
-    Index(['var1', 'var2', 'var3'], dtype='object')
+        layers: Noneå
+    >>> outer.var_names.astype("string")
+    Index(['var1', 'var2', 'var3'], dtype='string')
     >>> outer.to_df()  # Sparse arrays are padded with zeroes by default
         var1  var2  var3
     s1     0     1     0
@@ -1621,11 +1643,11 @@ def concat(  # noqa: PLR0912, PLR0913, PLR0915
 
     # Combining indexes
     concat_indices = pd.concat(
-        [pd.Series(axis_indices(a, axis=axis)) for a in adatas], ignore_index=True
+        [axis_indices(a, axis=axis).to_series() for a in adatas], ignore_index=True
     )
     if index_unique is not None:
         concat_indices = concat_indices.str.cat(
-            _map_cat_to_str(label_col), sep=index_unique
+            label_col.map(str, na_action="ignore"), sep=index_unique
         )
     concat_indices = pd.Index(concat_indices)
 
@@ -1681,6 +1703,9 @@ def concat(  # noqa: PLR0912, PLR0913, PLR0915
                 alt_annotations, use_only_object_dtype=True
             )
         )
+        for a in annotations_with_only_dask:
+            if a.true_index_dim != a.index_dim:
+                a.index = a.true_index
         annotations_with_only_dask = [
             a.ds.rename({a.true_index_dim: "merge_index"})
             for a in annotations_with_only_dask
@@ -1688,7 +1713,6 @@ def concat(  # noqa: PLR0912, PLR0913, PLR0915
         alt_annot = Dataset2D(
             xr.merge(annotations_with_only_dask, join=join, compat="override")
         )
-        alt_annot.true_index_dim = "merge_index"
 
     if join == "inner":
         concat_aligned_mapping = inner_concat_aligned_mapping
@@ -1728,15 +1752,10 @@ def concat(  # noqa: PLR0912, PLR0913, PLR0915
             for r, a in zip(reindexers, adatas, strict=True)
         ],
     )
-    alt_pairwise = merge(
-        [
-            {
-                k: r(r(v, axis=0), axis=1)
-                for k, v in getattr(a, f"{alt_axis_name}p").items()
-            }
-            for r, a in zip(reindexers, adatas, strict=True)
-        ]
-    )
+    alt_pairwise = merge([
+        {k: r(r(v, axis=0), axis=1) for k, v in getattr(a, f"{alt_axis_name}p").items()}
+        for r, a in zip(reindexers, adatas, strict=True)
+    ])
     uns = uns_merge([a.uns for a in adatas])
 
     raw = None
@@ -1765,16 +1784,14 @@ def concat(  # noqa: PLR0912, PLR0913, PLR0915
             "not concatenating `.raw` attributes."
         )
         warn(msg, UserWarning, stacklevel=2)
-    return AnnData(
-        **{
-            "layers": layers,
-            axis_name: concat_annot,
-            alt_axis_name: alt_annot,
-            f"{axis_name}m": concat_mapping,
-            f"{alt_axis_name}m": alt_mapping,
-            f"{axis_name}p": concat_pairwise,
-            f"{alt_axis_name}p": alt_pairwise,
-            "uns": uns,
-            "raw": raw,
-        }
-    )
+    return AnnData(**{
+        "layers": layers,
+        axis_name: concat_annot,
+        alt_axis_name: alt_annot,
+        f"{axis_name}m": concat_mapping,
+        f"{alt_axis_name}m": alt_mapping,
+        f"{axis_name}p": concat_pairwise,
+        f"{alt_axis_name}p": alt_pairwise,
+        "uns": uns,
+        "raw": raw,
+    })
