@@ -1,19 +1,18 @@
 from __future__ import annotations
 
-import warnings
 from collections.abc import Mapping
 from copy import copy
 from functools import partial
 from itertools import product
 from types import MappingProxyType
 from typing import TYPE_CHECKING
-from warnings import warn
 
 import h5py
 import numpy as np
 import pandas as pd
-from packaging.version import Version
+from numcodecs import VLenUTF8
 from scipy import sparse
+from zarr.core.dtype import VariableLengthUTF8
 
 import anndata as ad
 from anndata import AnnData, Raw
@@ -21,10 +20,9 @@ from anndata._core import views
 from anndata._core.index import _normalize_indices
 from anndata._core.merge import intersect_keys
 from anndata._core.sparse_dataset import _CSCDataset, _CSRDataset, sparse_dataset
-from anndata._io.utils import H5PY_V3, check_key, zero_dim_array_as_scalar
+from anndata._io.utils import check_key, zero_dim_array_as_scalar
 from anndata._warnings import OldFormatWarning
 from anndata.compat import (
-    NULLABLE_NUMPY_STRING_TYPE,
     AwkArray,
     CupyArray,
     CupyCSCMatrix,
@@ -42,7 +40,8 @@ from anndata.compat import (
 )
 
 from ..._settings import settings
-from ...compat import is_zarr_v2
+from ...compat import PANDAS_STRING_ARRAY_TYPES, PANDAS_SUPPORTS_NA_VALUE
+from ...utils import warn
 from .registry import _REGISTRY, IOSpec, read_elem, read_elem_partial
 
 if TYPE_CHECKING:
@@ -53,9 +52,9 @@ if TYPE_CHECKING:
     from numpy import typing as npt
     from numpy.typing import NDArray
 
-    from anndata._types import ArrayStorageType, GroupStorageType
+    from anndata._types import _ArrayStorageType, _GroupStorageType
     from anndata.compat import CSArray, CSMatrix
-    from anndata.typing import AxisStorable, InMemoryArrayOrScalarType
+    from anndata.typing import AxisStorable, _InMemoryArrayOrScalarType
 
     from .registry import Reader, Writer
 
@@ -95,9 +94,27 @@ GLOBAL_LOCK = Lock()
 #     return False
 
 
-def zarr_v3_compressor_compat(dataset_kwargs) -> dict:
-    if not is_zarr_v2() and (compressor := dataset_kwargs.pop("compressor", None)):
-        dataset_kwargs["compressors"] = compressor
+def zarr_v3_compressor_compat(dataset_kwargs: dict) -> dict:
+    """Handle mismatch between our compressor kwarg and :func:`zarr.create_array` in v3's `compressors` arg
+    See https://zarr.readthedocs.io/en/stable/api/zarr/create/#zarr.create_array
+
+    Parameters
+    ----------
+    dataset_kwarg
+        The kwarg dict potentially containing "compressor"
+
+    Returns
+    -------
+        The kwarg dict with "compressor" moved to "compressors" if zarr v3 is in use.
+    """
+    if "compressor" in dataset_kwargs:
+        dataset_kwargs["compressors"] = dataset_kwargs.pop("compressor")
+    return dataset_kwargs
+
+
+def zarr_v3_sharding(dataset_kwargs) -> dict:
+    if "shards" not in dataset_kwargs and ad.settings.auto_shard_zarr_v3:
+        dataset_kwargs = {**dataset_kwargs, "shards": "auto"}
     return dataset_kwargs
 
 
@@ -135,14 +152,11 @@ def _to_cpu_mem_wrapper(write_func):
 @_REGISTRY.register_read(H5Array, IOSpec("", ""))
 def read_basic(
     elem: H5File | H5Group | H5Array, *, _reader: Reader
-) -> dict[str, InMemoryArrayOrScalarType] | npt.NDArray | CSMatrix | CSArray:
+) -> dict[str, _InMemoryArrayOrScalarType] | npt.NDArray | CSMatrix | CSArray:
     from anndata._io import h5ad
 
-    warn(
-        f"Element '{elem.name}' was written without encoding metadata.",
-        OldFormatWarning,
-        stacklevel=3,
-    )
+    msg = f"Element '{elem.name}' was written without encoding metadata."
+    warn(msg, OldFormatWarning)
 
     if isinstance(elem, Mapping):
         # Backwards compat sparse arrays
@@ -157,14 +171,11 @@ def read_basic(
 @_REGISTRY.register_read(ZarrArray, IOSpec("", ""))
 def read_basic_zarr(
     elem: ZarrGroup | ZarrArray, *, _reader: Reader
-) -> dict[str, InMemoryArrayOrScalarType] | npt.NDArray | CSMatrix | CSArray:
+) -> dict[str, _InMemoryArrayOrScalarType] | npt.NDArray | CSMatrix | CSArray:
     from anndata._io import zarr
 
-    warn(
-        f"Element '{elem.name}' was written without encoding metadata.",
-        OldFormatWarning,
-        stacklevel=3,
-    )
+    msg = f"Element '{elem.name}' was written without encoding metadata."
+    warn(msg, OldFormatWarning)
     if isinstance(elem, ZarrGroup):
         # Backwards compat sparse arrays
         if "h5sparse_format" in elem.attrs:
@@ -267,7 +278,7 @@ def _read_partial(group, *, items=None, indices=(slice(None), slice(None))):
 @_REGISTRY.register_write(ZarrGroup, AnnData, IOSpec("anndata", "0.1.0"))
 @_REGISTRY.register_write(H5Group, AnnData, IOSpec("anndata", "0.1.0"))
 def write_anndata(
-    f: GroupStorageType,
+    f: _GroupStorageType,
     k: str,
     adata: AnnData,
     *,
@@ -299,7 +310,7 @@ def write_anndata(
 @_REGISTRY.register_read(H5File, IOSpec("raw", "0.1.0"))
 @_REGISTRY.register_read(ZarrGroup, IOSpec("anndata", "0.1.0"))
 @_REGISTRY.register_read(ZarrGroup, IOSpec("raw", "0.1.0"))
-def read_anndata(elem: GroupStorageType | H5File, *, _reader: Reader) -> AnnData:
+def read_anndata(elem: _GroupStorageType | H5File, *, _reader: Reader) -> AnnData:
     d = {}
     for k in [
         "X",
@@ -321,7 +332,7 @@ def read_anndata(elem: GroupStorageType | H5File, *, _reader: Reader) -> AnnData
 @_REGISTRY.register_write(H5Group, Raw, IOSpec("raw", "0.1.0"))
 @_REGISTRY.register_write(ZarrGroup, Raw, IOSpec("raw", "0.1.0"))
 def write_raw(
-    f: GroupStorageType,
+    f: _GroupStorageType,
     k: str,
     raw: Raw,
     *,
@@ -355,15 +366,9 @@ def write_null_h5py(f, k, _v, _writer, dataset_kwargs=MappingProxyType({})):
 def write_null_zarr(f, k, _v, _writer, dataset_kwargs=MappingProxyType({})):
     dataset_kwargs = _remove_scalar_compression_args(dataset_kwargs)
     # zarr has no first-class null dataset
-    if is_zarr_v2():
-        import zarr
-
-        # zarr has no first-class null dataset
-        f.create_dataset(k, data=zarr.empty(()), **dataset_kwargs)
-    else:
-        # TODO: why is this not actually storing the empty info with a f.empty call?
-        # It fails complaining that k doesn't exist when updating the attributes.
-        f.create_array(k, shape=(), dtype="bool")
+    # TODO: why is this not actually storing the empty info with a f.empty call?
+    # It fails complaining that k doesn't exist when updating the attributes.
+    f.create_array(k, shape=(), dtype="bool")
 
 
 ############
@@ -373,14 +378,16 @@ def write_null_zarr(f, k, _v, _writer, dataset_kwargs=MappingProxyType({})):
 
 @_REGISTRY.register_read(H5Group, IOSpec("dict", "0.1.0"))
 @_REGISTRY.register_read(ZarrGroup, IOSpec("dict", "0.1.0"))
-def read_mapping(elem: GroupStorageType, *, _reader: Reader) -> dict[str, AxisStorable]:
+def read_mapping(
+    elem: _GroupStorageType, *, _reader: Reader
+) -> dict[str, AxisStorable]:
     return {k: _reader.read_elem(v) for k, v in dict(elem).items()}
 
 
 @_REGISTRY.register_write(H5Group, dict, IOSpec("dict", "0.1.0"))
 @_REGISTRY.register_write(ZarrGroup, dict, IOSpec("dict", "0.1.0"))
 def write_mapping(
-    f: GroupStorageType,
+    f: _GroupStorageType,
     k: str,
     v: dict[str, AxisStorable],
     *,
@@ -400,7 +407,7 @@ def write_mapping(
 @_REGISTRY.register_write(H5Group, list, IOSpec("array", "0.2.0"))
 @_REGISTRY.register_write(ZarrGroup, list, IOSpec("array", "0.2.0"))
 def write_list(
-    f: GroupStorageType,
+    f: _GroupStorageType,
     k: str,
     elem: list[AxisStorable],
     *,
@@ -422,7 +429,7 @@ def write_list(
 @_REGISTRY.register_write(ZarrGroup, H5Array, IOSpec("array", "0.2.0"))
 @zero_dim_array_as_scalar
 def write_basic(
-    f: GroupStorageType,
+    f: _GroupStorageType,
     k: str,
     elem: views.ArrayView | np.ndarray | h5py.Dataset | np.ma.MaskedArray | ZarrArray,
     *,
@@ -432,10 +439,11 @@ def write_basic(
     """Write methods which underlying library handles natively."""
     dataset_kwargs = dataset_kwargs.copy()
     dtype = dataset_kwargs.pop("dtype", elem.dtype)
-    if isinstance(f, H5Group) or is_zarr_v2():
+    if isinstance(f, H5Group):
         f.create_dataset(k, data=elem, shape=elem.shape, dtype=dtype, **dataset_kwargs)
     else:
         dataset_kwargs = zarr_v3_compressor_compat(dataset_kwargs)
+        dataset_kwargs = zarr_v3_sharding(dataset_kwargs)
         f.create_array(k, shape=elem.shape, dtype=dtype, **dataset_kwargs)
         # see https://github.com/zarr-developers/zarr-python/discussions/2712
         if isinstance(elem, ZarrArray | H5Array):
@@ -445,7 +453,7 @@ def write_basic(
 
 
 def _iter_chunks_for_copy(
-    elem: ArrayStorageType, dest: ArrayStorageType
+    elem: _ArrayStorageType, dest: _ArrayStorageType
 ) -> Iterator[slice | tuple[list[slice]]]:
     """
     Returns an iterator of tuples of slices for copying chunks from `elem` to `dest`.
@@ -470,7 +478,7 @@ def _iter_chunks_for_copy(
 def write_chunked_dense_array_to_group(
     f: H5Group,
     k: str,
-    elem: ArrayStorageType,
+    elem: _ArrayStorageType,
     *,
     _writer: Writer,
     dataset_kwargs: Mapping[str, Any] = MappingProxyType({}),
@@ -499,8 +507,10 @@ _REGISTRY.register_write(ZarrGroup, CupyArray, IOSpec("array", "0.2.0"))(
 
 @_REGISTRY.register_write(ZarrGroup, views.DaskArrayView, IOSpec("array", "0.2.0"))
 @_REGISTRY.register_write(ZarrGroup, DaskArray, IOSpec("array", "0.2.0"))
-def write_basic_dask_zarr(
-    f: ZarrGroup,
+@_REGISTRY.register_write(H5Group, views.DaskArrayView, IOSpec("array", "0.2.0"))
+@_REGISTRY.register_write(H5Group, DaskArray, IOSpec("array", "0.2.0"))
+def write_basic_dask_dask_dense(
+    f: ZarrGroup | H5Group,
     k: str,
     elem: DaskArray,
     *,
@@ -510,41 +520,21 @@ def write_basic_dask_zarr(
     import dask.array as da
 
     dataset_kwargs = dataset_kwargs.copy()
-    dataset_kwargs = zarr_v3_compressor_compat(dataset_kwargs)
-    if is_zarr_v2():
+    is_h5 = isinstance(f, H5Group)
+    if not is_h5:
+        dataset_kwargs = zarr_v3_compressor_compat(dataset_kwargs)
+        dataset_kwargs = zarr_v3_sharding(dataset_kwargs)
+    if is_h5:
         g = f.require_dataset(k, shape=elem.shape, dtype=elem.dtype, **dataset_kwargs)
     else:
         g = f.require_array(k, shape=elem.shape, dtype=elem.dtype, **dataset_kwargs)
-    da.store(elem, g, lock=GLOBAL_LOCK)
-
-
-# Adding this separately because h5py isn't serializable
-# https://github.com/pydata/xarray/issues/4242
-@_REGISTRY.register_write(H5Group, views.DaskArrayView, IOSpec("array", "0.2.0"))
-@_REGISTRY.register_write(H5Group, DaskArray, IOSpec("array", "0.2.0"))
-def write_basic_dask_h5(
-    f: H5Group,
-    k: str,
-    elem: DaskArray,
-    *,
-    _writer: Writer,
-    dataset_kwargs: Mapping[str, Any] = MappingProxyType({}),
-):
-    import dask.array as da
-    import dask.config as dc
-
-    if dc.get("scheduler", None) == "dask.distributed":
-        msg = "Cannot write dask arrays to hdf5 when using distributed scheduler"
-        raise ValueError(msg)
-
-    g = f.require_dataset(k, shape=elem.shape, dtype=elem.dtype, **dataset_kwargs)
-    da.store(elem, g)
+    da.store(elem, g, scheduler="threads")
 
 
 @_REGISTRY.register_read(H5Array, IOSpec("array", "0.2.0"))
 @_REGISTRY.register_read(ZarrArray, IOSpec("array", "0.2.0"))
 @_REGISTRY.register_read(ZarrArray, IOSpec("string-array", "0.2.0"))
-def read_array(elem: ArrayStorageType, *, _reader: Reader) -> npt.NDArray:
+def read_array(elem: _ArrayStorageType, *, _reader: Reader) -> npt.NDArray:
     return elem[()]
 
 
@@ -611,43 +601,22 @@ def write_vlen_string_array_zarr(
     _writer: Writer,
     dataset_kwargs: Mapping[str, Any] = MappingProxyType({}),
 ):
-    if is_zarr_v2():
-        import numcodecs
-
-        if Version(numcodecs.__version__) < Version("0.13"):
-            msg = "Old numcodecs version detected. Please update for improved performance and stability."
-            warnings.warn(msg, UserWarning, stacklevel=2)
-            # Workaround for https://github.com/zarr-developers/numcodecs/issues/514
-            if hasattr(elem, "flags") and not elem.flags.writeable:
-                elem = elem.copy()
-
-        f.create_dataset(
-            k,
-            shape=elem.shape,
-            dtype=object,
-            object_codec=numcodecs.VLenUTF8(),
-            **dataset_kwargs,
-        )
-        f[k][:] = elem
-    else:
-        from numcodecs import VLenUTF8
-        from zarr.core.dtype import VariableLengthUTF8
-
-        dataset_kwargs = dataset_kwargs.copy()
-        dataset_kwargs = zarr_v3_compressor_compat(dataset_kwargs)
-        dtype = VariableLengthUTF8()
-        filters, fill_value = None, None
-        if f.metadata.zarr_format == 2:
-            filters, fill_value = [VLenUTF8()], ""
-        f.create_array(
-            k,
-            shape=elem.shape,
-            dtype=dtype,
-            filters=filters,
-            fill_value=fill_value,
-            **dataset_kwargs,
-        )
-        f[k][:] = elem
+    dataset_kwargs = dataset_kwargs.copy()
+    dataset_kwargs = zarr_v3_compressor_compat(dataset_kwargs)
+    dtype = VariableLengthUTF8()
+    filters, fill_value = None, None
+    if f.metadata.zarr_format == 2:
+        filters, fill_value = [VLenUTF8()], ""
+    dataset_kwargs = zarr_v3_sharding(dataset_kwargs)
+    f.create_array(
+        k,
+        shape=elem.shape,
+        dtype=dtype,
+        filters=filters,
+        fill_value=fill_value,
+        **dataset_kwargs,
+    )
+    f[k][:] = elem
 
 
 ###############
@@ -668,12 +637,13 @@ def _to_hdf5_vlen_strings(value: np.ndarray) -> np.ndarray:
 
 @_REGISTRY.register_read(H5Array, IOSpec("rec-array", "0.2.0"))
 @_REGISTRY.register_read(ZarrArray, IOSpec("rec-array", "0.2.0"))
-def read_recarray(d: ArrayStorageType, *, _reader: Reader) -> np.recarray | npt.NDArray:
+def read_recarray(
+    d: _ArrayStorageType, *, _reader: Reader
+) -> np.recarray | npt.NDArray:
     value = d[()]
-    dtype = value.dtype
-    value = _from_fixed_length_strings(value)
-    if H5PY_V3:
-        value = _decode_structured_array(value, dtype=dtype)
+    value = _decode_structured_array(
+        _from_fixed_length_strings(value), dtype=value.dtype
+    )
     return value
 
 
@@ -703,13 +673,13 @@ def write_recarray_zarr(
     from anndata.compat import _to_fixed_length_strings
 
     elem = _to_fixed_length_strings(elem)
-    if is_zarr_v2():
-        f.create_dataset(k, data=elem, shape=elem.shape, **dataset_kwargs)
-    else:
-        dataset_kwargs = dataset_kwargs.copy()
-        dataset_kwargs = zarr_v3_compressor_compat(dataset_kwargs)
-        f.create_array(k, shape=elem.shape, dtype=elem.dtype, **dataset_kwargs)
-        f[k][...] = elem
+    dataset_kwargs = dataset_kwargs.copy()
+    dataset_kwargs = zarr_v3_compressor_compat(dataset_kwargs)
+    # https://github.com/zarr-developers/zarr-python/issues/3546
+    # if "shards" not in dataset_kwargs and ad.settings.auto_shard_zarr_v3:
+    #     dataset_kwargs = {**dataset_kwargs, "shards": "auto"}
+    f.create_array(k, shape=elem.shape, dtype=elem.dtype, **dataset_kwargs)
+    f[k][...] = elem
 
 
 #################
@@ -718,7 +688,7 @@ def write_recarray_zarr(
 
 
 def write_sparse_compressed(
-    f: GroupStorageType,
+    f: _GroupStorageType,
     key: str,
     value: CSMatrix | CSArray,
     *,
@@ -739,11 +709,30 @@ def write_sparse_compressed(
     for attr_name in ["data", "indices", "indptr"]:
         attr = getattr(value, attr_name)
         dtype = indptr_dtype if attr_name == "indptr" else attr.dtype
-        if isinstance(f, H5Group) or is_zarr_v2():
+        if (
+            attr_name == "indices"
+            and settings.write_csr_csc_indices_with_min_possible_dtype
+        ):
+            # np.min_scalar_type can return things like np.ulonglong which zarr doesn't understand
+            # and I find this clearer as to what the result type is i.e., unsigned or signed.
+            # For example `np.iinfo(np.uint16).max + 1` could be either `uint32` or `int32`,
+            # and there's nothing in numpy's docs disallowing this output to change.
+            if (minor_axis_size := value.shape[value.format == "csr"]) <= np.iinfo(
+                np.uint8
+            ).max:
+                dtype = np.dtype("uint8")
+            elif minor_axis_size <= np.iinfo(np.uint16).max:
+                dtype = np.dtype("uint16")
+            elif minor_axis_size <= np.iinfo(np.uint32).max:
+                dtype = np.dtype("uint32")
+            elif minor_axis_size <= np.iinfo(np.uint64).max:
+                dtype = np.dtype("uint64")
+        if isinstance(f, H5Group):
             g.create_dataset(
                 attr_name, data=attr, shape=attr.shape, dtype=dtype, **dataset_kwargs
             )
         else:
+            dataset_kwargs = zarr_v3_sharding(dataset_kwargs)
             arr = g.create_array(
                 attr_name, shape=attr.shape, dtype=dtype, **dataset_kwargs
             )
@@ -785,12 +774,12 @@ for store_type, (cls, spec, func) in product(
     _REGISTRY.register_write(store_type, cls, spec)(func)
 
 
-@_REGISTRY.register_write(H5Group, _CSRDataset, IOSpec("", "0.1.0"))
-@_REGISTRY.register_write(H5Group, _CSCDataset, IOSpec("", "0.1.0"))
-@_REGISTRY.register_write(ZarrGroup, _CSRDataset, IOSpec("", "0.1.0"))
-@_REGISTRY.register_write(ZarrGroup, _CSCDataset, IOSpec("", "0.1.0"))
+@_REGISTRY.register_write(H5Group, _CSRDataset, IOSpec("csr_matrix", "0.1.0"))
+@_REGISTRY.register_write(H5Group, _CSCDataset, IOSpec("csc_matrix", "0.1.0"))
+@_REGISTRY.register_write(ZarrGroup, _CSRDataset, IOSpec("csr_matrix", "0.1.0"))
+@_REGISTRY.register_write(ZarrGroup, _CSCDataset, IOSpec("csc_matrix", "0.1.0"))
 def write_sparse_dataset(
-    f: GroupStorageType,
+    f: _GroupStorageType,
     k: str,
     elem: _CSCDataset | _CSRDataset,
     *,
@@ -805,9 +794,6 @@ def write_sparse_dataset(
         fmt=elem.format,
         dataset_kwargs=dataset_kwargs,
     )
-    # TODO: Cleaner way to do this
-    f[k].attrs["encoding-type"] = f"{elem.format}_matrix"
-    f[k].attrs["encoding-version"] = "0.1.0"
 
 
 def write_cupy_dask(f, k, elem, _writer, dataset_kwargs=MappingProxyType({})):
@@ -820,7 +806,7 @@ def write_cupy_dask(f, k, elem, _writer, dataset_kwargs=MappingProxyType({})):
 
 
 def write_dask_sparse(
-    f: GroupStorageType,
+    f: _GroupStorageType,
     k: str,
     elem: DaskArray,
     *,
@@ -881,6 +867,8 @@ for array_type, group_type in product(
     for scipy_sparse_type, spec in [
         (sparse.csr_matrix, IOSpec("csr_matrix", "0.1.0")),
         (sparse.csc_matrix, IOSpec("csc_matrix", "0.1.0")),
+        (sparse.csr_array, IOSpec("csr_matrix", "0.1.0")),
+        (sparse.csc_array, IOSpec("csc_matrix", "0.1.0")),
     ]:
         _REGISTRY.register_write(group_type, (array_type, scipy_sparse_type), spec)(
             write_dask_sparse
@@ -891,7 +879,7 @@ for array_type, group_type in product(
 @_REGISTRY.register_read(H5Group, IOSpec("csr_matrix", "0.1.0"))
 @_REGISTRY.register_read(ZarrGroup, IOSpec("csc_matrix", "0.1.0"))
 @_REGISTRY.register_read(ZarrGroup, IOSpec("csr_matrix", "0.1.0"))
-def read_sparse(elem: GroupStorageType, *, _reader: Reader) -> CSMatrix | CSArray:
+def read_sparse(elem: _GroupStorageType, *, _reader: Reader) -> CSMatrix | CSArray:
     return sparse_dataset(elem).to_memory()
 
 
@@ -917,7 +905,7 @@ def read_sparse_partial(elem, *, items=None, indices=(slice(None), slice(None)))
     ZarrGroup, views.AwkwardArrayView, IOSpec("awkward-array", "0.1.0")
 )
 def write_awkward(
-    f: GroupStorageType,
+    f: _GroupStorageType,
     k: str,
     v: views.AwkwardArrayView | AwkArray,
     *,
@@ -940,7 +928,7 @@ def write_awkward(
 
 @_REGISTRY.register_read(H5Group, IOSpec("awkward-array", "0.1.0"))
 @_REGISTRY.register_read(ZarrGroup, IOSpec("awkward-array", "0.1.0"))
-def read_awkward(elem: GroupStorageType, *, _reader: Reader) -> AwkArray:
+def read_awkward(elem: _GroupStorageType, *, _reader: Reader) -> AwkArray:
     from anndata.compat import awkward as ak
 
     form = _read_attr(elem.attrs, "form")
@@ -960,7 +948,7 @@ def read_awkward(elem: GroupStorageType, *, _reader: Reader) -> AwkArray:
 @_REGISTRY.register_write(ZarrGroup, views.DataFrameView, IOSpec("dataframe", "0.2.0"))
 @_REGISTRY.register_write(ZarrGroup, pd.DataFrame, IOSpec("dataframe", "0.2.0"))
 def write_dataframe(
-    f: GroupStorageType,
+    f: _GroupStorageType,
     key: str,
     df: views.DataFrameView | pd.DataFrame,
     *,
@@ -1010,7 +998,7 @@ def write_dataframe(
 
 @_REGISTRY.register_read(H5Group, IOSpec("dataframe", "0.2.0"))
 @_REGISTRY.register_read(ZarrGroup, IOSpec("dataframe", "0.2.0"))
-def read_dataframe(elem: GroupStorageType, *, _reader: Reader) -> pd.DataFrame:
+def read_dataframe(elem: _GroupStorageType, *, _reader: Reader) -> pd.DataFrame:
     columns = list(_read_attr(elem.attrs, "column-order"))
     idx_key = _read_attr(elem.attrs, "_index")
     df = pd.DataFrame(
@@ -1051,7 +1039,7 @@ def read_dataframe_partial(
 
 @_REGISTRY.register_read(H5Group, IOSpec("dataframe", "0.1.0"))
 @_REGISTRY.register_read(ZarrGroup, IOSpec("dataframe", "0.1.0"))
-def read_dataframe_0_1_0(elem: GroupStorageType, *, _reader: Reader) -> pd.DataFrame:
+def read_dataframe_0_1_0(elem: _GroupStorageType, *, _reader: Reader) -> pd.DataFrame:
     columns = _read_attr(elem.attrs, "column-order")
     idx_key = _read_attr(elem.attrs, "_index")
     df = pd.DataFrame(
@@ -1101,7 +1089,7 @@ def read_partial_dataframe_0_1_0(
 @_REGISTRY.register_write(H5Group, pd.Categorical, IOSpec("categorical", "0.2.0"))
 @_REGISTRY.register_write(ZarrGroup, pd.Categorical, IOSpec("categorical", "0.2.0"))
 def write_categorical(
-    f: GroupStorageType,
+    f: _GroupStorageType,
     k: str,
     v: pd.Categorical,
     *,
@@ -1113,13 +1101,16 @@ def write_categorical(
 
     _writer.write_elem(g, "codes", v.codes, dataset_kwargs=dataset_kwargs)
     _writer.write_elem(
-        g, "categories", v.categories._values, dataset_kwargs=dataset_kwargs
+        g,
+        "categories",
+        v.categories.to_numpy(),
+        dataset_kwargs=dataset_kwargs,
     )
 
 
 @_REGISTRY.register_read(H5Group, IOSpec("categorical", "0.2.0"))
 @_REGISTRY.register_read(ZarrGroup, IOSpec("categorical", "0.2.0"))
-def read_categorical(elem: GroupStorageType, *, _reader: Reader) -> pd.Categorical:
+def read_categorical(elem: _GroupStorageType, *, _reader: Reader) -> pd.Categorical:
     return pd.Categorical.from_codes(
         codes=_reader.read_elem(elem["codes"]),
         categories=_reader.read_elem(elem["categories"]),
@@ -1154,43 +1145,72 @@ def read_partial_categorical(elem, *, items=None, indices=(slice(None),)):
 @_REGISTRY.register_write(
     ZarrGroup, pd.arrays.BooleanArray, IOSpec("nullable-boolean", "0.1.0")
 )
-@_REGISTRY.register_write(
-    H5Group, pd.arrays.StringArray, IOSpec("nullable-string-array", "0.1.0")
-)
-@_REGISTRY.register_write(
-    ZarrGroup, pd.arrays.StringArray, IOSpec("nullable-string-array", "0.1.0")
-)
 def write_nullable(
-    f: GroupStorageType,
+    f: _GroupStorageType,
     k: str,
-    v: pd.arrays.IntegerArray | pd.arrays.BooleanArray | pd.arrays.StringArray,
+    v: pd.arrays.IntegerArray
+    | pd.arrays.BooleanArray
+    | pd.arrays.StringArray
+    | pd.arrays.ArrowStringArray,
     *,
     _writer: Writer,
     dataset_kwargs: Mapping[str, Any] = MappingProxyType({}),
-):
-    if (
-        isinstance(v, pd.arrays.StringArray)
-        and not settings.allow_write_nullable_strings
-    ):
-        msg = (
-            "`anndata.settings.allow_write_nullable_strings` is False, "
-            "because writing of `pd.arrays.StringArray` is new "
-            "and not supported in anndata < 0.11, still use by many people. "
-            "Opt-in to writing these arrays by toggling the setting to True."
-        )
-        raise RuntimeError(msg)
+) -> None:
+    if isinstance(v, pd.arrays.StringArray | pd.arrays.ArrowStringArray):
+        # if explicitly set to `False`, we try writing them as non-nullable string array
+        if settings.allow_write_nullable_strings is False:
+            if v.isna().any():
+                msg = (
+                    "Cannot write `pd.arrays.[Arrow]StringArray` with missing values "
+                    "when `anndata.settings.allow_write_nullable_strings` is set to False. "
+                    "Opt-in to writing these arrays by toggling it to True or remove missing values."
+                )
+                raise ValueError(msg)
+            # This works since there are no missing values. Otherwise, we’d need to be very careful.
+            # TODO: check if the cost of creating a numpy "T" string array here would amortize when writing
+            _writer.write_elem(f, k, np.asarray(v), dataset_kwargs=dataset_kwargs)
+            return
+
+        # if implicitly set to `False`, we error out
+        if (
+            settings.allow_write_nullable_strings is None
+            and not pd.options.future.infer_string
+        ):
+            msg = (
+                "`anndata.settings.allow_write_nullable_strings` is None and "
+                "`pd.options.future.infer_string` is False. "
+                "Opt-in to writing these arrays by toggling either setting to True, "
+                "or make `anndata` attempt to write the non-nullable format supported by "
+                "anndata < 0.11 by setting `anndata.settings.allow_write_nullable_strings` to False."
+            )
+            raise RuntimeError(msg)
+
     g = f.require_group(k)
+    if isinstance(v, pd.arrays.StringArray | pd.arrays.ArrowStringArray):
+        if v.dtype.na_value is pd.NA:
+            g.attrs["na-value"] = "NA"
+        elif np.isnan(v.dtype.na_value):
+            g.attrs["na-value"] = "NaN"
+        else:
+            msg = f"Cannot write {v.dtype.na_value} as na_value for pandas StringArray"
+            raise ValueError(msg)
     values = (
         v.to_numpy(na_value="")
-        if isinstance(v, pd.arrays.StringArray)
+        if isinstance(v, pd.arrays.StringArray | pd.arrays.ArrowStringArray)
         else v.to_numpy(na_value=0, dtype=v.dtype.numpy_dtype)
     )
     _writer.write_elem(g, "values", values, dataset_kwargs=dataset_kwargs)
     _writer.write_elem(g, "mask", v.isna(), dataset_kwargs=dataset_kwargs)
 
 
+for store_type, array_type in product([H5Group, ZarrGroup], PANDAS_STRING_ARRAY_TYPES):
+    _REGISTRY.register_write(
+        store_type, array_type, IOSpec("nullable-string-array", "0.1.0")
+    )(write_nullable)
+
+
 def _read_nullable(
-    elem: GroupStorageType,
+    elem: _GroupStorageType,
     *,
     _reader: Reader,
     # BaseMaskedArray
@@ -1202,18 +1222,6 @@ def _read_nullable(
         _reader.read_elem(elem["values"]),
         mask=_reader.read_elem(elem["mask"]),
     )
-
-
-def _string_array(
-    values: np.ndarray, mask: np.ndarray
-) -> pd.api.extensions.ExtensionArray:
-    """Construct a string array from values and mask."""
-    arr = pd.array(
-        values.astype(NULLABLE_NUMPY_STRING_TYPE),
-        dtype=pd.StringDtype(),
-    )
-    arr[mask] = pd.NA
-    return arr
 
 
 _REGISTRY.register_read(H5Group, IOSpec("nullable-integer", "0.1.0"))(
@@ -1230,12 +1238,30 @@ _REGISTRY.register_read(ZarrGroup, IOSpec("nullable-boolean", "0.1.0"))(
     read_nullable_boolean
 )
 
-_REGISTRY.register_read(H5Group, IOSpec("nullable-string-array", "0.1.0"))(
-    read_nullable_string := partial(_read_nullable, array_type=_string_array)
-)
-_REGISTRY.register_read(ZarrGroup, IOSpec("nullable-string-array", "0.1.0"))(
-    read_nullable_string
-)
+
+@_REGISTRY.register_read(H5Group, IOSpec("nullable-string-array", "0.1.0"))
+@_REGISTRY.register_read(ZarrGroup, IOSpec("nullable-string-array", "0.1.0"))
+def _read_nullable_string(
+    elem: _GroupStorageType, *, _reader: Reader
+) -> pd.api.extensions.ExtensionArray:
+    values = _reader.read_elem(elem["values"])
+    mask = _reader.read_elem(elem["mask"])
+    dtype = (
+        pd.StringDtype(
+            na_value=np.nan
+            if _read_attr(elem.attrs, "na-value", default="NA") == "NaN"
+            else pd.NA
+        )
+        if PANDAS_SUPPORTS_NA_VALUE
+        else pd.StringDtype()
+    )
+
+    arr = pd.array(
+        values.astype(np.dtypes.StringDType(na_object=dtype.na_value)),  # TODO: why?
+        dtype=dtype,
+    )
+    arr[mask] = pd.NA
+    return arr
 
 
 ###########
@@ -1245,7 +1271,7 @@ _REGISTRY.register_read(ZarrGroup, IOSpec("nullable-string-array", "0.1.0"))(
 
 @_REGISTRY.register_read(H5Array, IOSpec("numeric-scalar", "0.2.0"))
 @_REGISTRY.register_read(ZarrArray, IOSpec("numeric-scalar", "0.2.0"))
-def read_scalar(elem: ArrayStorageType, *, _reader: Reader) -> np.number:
+def read_scalar(elem: _ArrayStorageType, *, _reader: Reader) -> np.number:
     # TODO: `item` ensures the return is in fact a scalar (needed after zarr v3 which now returns a 1 elem array)
     # https://github.com/zarr-developers/zarr-python/issues/2713
     return elem[()].item()
@@ -1278,28 +1304,23 @@ def write_scalar_zarr(
     # these args are ignored in v2: https://zarr.readthedocs.io/en/v2.18.4/api/hierarchy.html#zarr.hierarchy.Group.create_dataset
     # and error out in v3
     dataset_kwargs = _remove_scalar_compression_args(dataset_kwargs)
-    if is_zarr_v2():
-        return f.create_dataset(key, data=np.array(value), shape=(), **dataset_kwargs)
-    else:
-        from numcodecs import VLenUTF8
-        from zarr.core.dtype import VariableLengthUTF8
 
-        match f.metadata.zarr_format, value:
-            case 2, str():
-                filters, dtype, fill_value = [VLenUTF8()], VariableLengthUTF8(), ""
-            case 3, str():
-                filters, dtype, fill_value = None, VariableLengthUTF8(), None
-            case _, _:
-                filters, dtype, fill_value = None, np.array(value).dtype, None
-        a = f.create_array(
-            key,
-            shape=(),
-            dtype=dtype,
-            filters=filters,
-            fill_value=fill_value,
-            **dataset_kwargs,
-        )
-        a[...] = np.array(value)
+    match f.metadata.zarr_format, value:
+        case 2, str():
+            filters, dtype, fill_value = [VLenUTF8()], VariableLengthUTF8(), ""
+        case 3, str():
+            filters, dtype, fill_value = None, VariableLengthUTF8(), None
+        case _, _:
+            filters, dtype, fill_value = None, np.array(value).dtype, None
+    a = f.create_array(
+        key,
+        shape=(),
+        dtype=dtype,
+        filters=filters,
+        fill_value=fill_value,
+        **dataset_kwargs,
+    )
+    a[...] = np.array(value)
 
 
 def write_hdf5_scalar(
