@@ -4,10 +4,11 @@ Defines some useful types for this library. Should probably be cleaned up before
 
 from __future__ import annotations
 
-from typing import TYPE_CHECKING, Literal, Protocol, TypeVar
+from typing import TYPE_CHECKING, Literal, Protocol
 
 from . import typing
 from .compat import H5Array, H5Group, ZarrArray, ZarrGroup
+from .utils import set_module
 
 if TYPE_CHECKING:
     from collections.abc import Mapping
@@ -22,50 +23,49 @@ if TYPE_CHECKING:
         Reader,
         Writer,
     )
+else:  # https://github.com/tox-dev/sphinx-autodoc-typehints/issues/580
+    type S = StorageType
+    type RWAble = typing.RWAble
 
 
 __all__ = [
-    "ArrayStorageType",
-    "GroupStorageType",
     "StorageType",
+    "_ArrayStorageType",
+    "_GroupStorageType",
     "_ReadInternal",
     "_ReadLazyInternal",
     "_WriteInternal",
 ]
 
-ArrayStorageType: TypeAlias = ZarrArray | H5Array
-GroupStorageType: TypeAlias = ZarrGroup | H5Group
-StorageType: TypeAlias = ArrayStorageType | GroupStorageType
+# These two are not public, so we don’t make them `type`s
+_ArrayStorageType: TypeAlias = ZarrArray | H5Array  # noqa: UP040
+_GroupStorageType: TypeAlias = ZarrGroup | H5Group  # noqa: UP040
 
-# NOTE: If you change these, be sure to update `autodoc_type_aliases` in docs/conf.py!
-RWAble_contra = TypeVar("RWAble_contra", bound=typing.RWAble, contravariant=True)
-RWAble_co = TypeVar("RWAble_co", bound=typing.RWAble, covariant=True)
-RWAble = TypeVar("RWAble", bound=typing.RWAble)
-
-S_co = TypeVar("S_co", covariant=True, bound=StorageType)
-S_contra = TypeVar("S_contra", contravariant=True, bound=StorageType)
+type StorageType = _ArrayStorageType | _GroupStorageType
 
 
+@set_module("anndata.experimental")
 class Dataset2DIlocIndexer(Protocol):
     def __getitem__(self, idx: Any) -> Dataset2D: ...
 
 
-class _ReadInternal(Protocol[S_contra, RWAble_co]):
-    def __call__(self, elem: S_contra, *, _reader: Reader) -> RWAble_co: ...
+class _ReadInternal[S: StorageType, RWAble: typing.RWAble](Protocol):
+    def __call__(self, elem: S, *, _reader: Reader) -> RWAble: ...
 
 
-class _ReadLazyInternal(Protocol[S_contra]):
+class _ReadLazyInternal[S: StorageType](Protocol):
     def __call__(
         self,
-        elem: S_contra,
+        elem: S,
         *,
         _reader: LazyReader,
         chunks: tuple[int, ...] | None = None,
     ) -> LazyDataStructures: ...
 
 
-class Read(Protocol[S_contra, RWAble_co]):
-    def __call__(self, elem: S_contra) -> RWAble_co:
+@set_module("anndata.experimental")
+class Read[S: StorageType, RWAble: typing.RWAble](Protocol):
+    def __call__(self, elem: S) -> RWAble:
         """Low-level reading function for an element.
 
         Parameters
@@ -79,9 +79,9 @@ class Read(Protocol[S_contra, RWAble_co]):
         ...
 
 
-class ReadLazy(Protocol[S_contra]):
+class ReadLazy[S](Protocol):
     def __call__(
-        self, elem: S_contra, *, chunks: tuple[int, ...] | None = None
+        self, elem: S, *, chunks: tuple[int, ...] | None = None
     ) -> LazyDataStructures:
         """Low-level reading function for a lazy element.
 
@@ -98,24 +98,25 @@ class ReadLazy(Protocol[S_contra]):
         ...
 
 
-class _WriteInternal(Protocol[RWAble_contra]):
+class _WriteInternal[RWAble: typing.RWAble](Protocol):
     def __call__(
         self,
         f: StorageType,
         k: str,
-        v: RWAble_contra,
+        v: RWAble,
         *,
         _writer: Writer,
         dataset_kwargs: Mapping[str, Any],
     ) -> None: ...
 
 
-class Write(Protocol[RWAble_contra]):
+@set_module("anndata.experimental")
+class Write[RWAble: typing.RWAble](Protocol):
     def __call__(
         self,
         f: StorageType,
         k: str,
-        v: RWAble_contra,
+        v: RWAble,
         *,
         dataset_kwargs: Mapping[str, Any],
     ) -> None:
@@ -130,16 +131,17 @@ class Write(Protocol[RWAble_contra]):
         v
             The element to write out.
         dataset_kwargs
-            Keyword arguments to be passed to a library-level io function, like `chunks` for :doc:`zarr:index`.
+            Keyword arguments to be passed to a library-level io function, like `chunks` for :mod:`zarr`.
         """
         ...
 
 
-class ReadCallback(Protocol[S_co, RWAble]):
+@set_module("anndata.experimental")
+class ReadCallback[S: StorageType, RWAble: typing.RWAble](Protocol):
     def __call__(
         self,
         /,
-        read_func: Read[S_co, RWAble],
+        read_func: Read[S, RWAble],
         elem_name: str,
         elem: StorageType,
         *,
@@ -166,7 +168,8 @@ class ReadCallback(Protocol[S_co, RWAble]):
         ...
 
 
-class WriteCallback(Protocol[RWAble]):
+@set_module("anndata.experimental")
+class WriteCallback[RWAble: typing.RWAble](Protocol):
     def __call__(
         self,
         /,
@@ -194,12 +197,12 @@ class WriteCallback(Protocol[RWAble]):
         iospec
             Internal AnnData encoding specification for the element.
         dataset_kwargs
-            Keyword arguments to be passed to a library-level io function, like `chunks` for :doc:`zarr:index`.
+            Keyword arguments to be passed to a library-level io function, like `chunks` for :mod:`zarr`.
         """
         ...
 
 
-AnnDataElem = Literal[
+type AnnDataElem = Literal[
     "obs",
     "var",
     "obsm",
@@ -212,4 +215,4 @@ AnnDataElem = Literal[
     "uns",
 ]
 
-Join_T = Literal["inner", "outer"]
+type Join_T = Literal["inner", "outer"]
