@@ -2,7 +2,7 @@
 
 from __future__ import annotations
 
-from pathlib import Path
+from typing import TYPE_CHECKING
 
 import joblib
 import numpy as np
@@ -10,7 +10,6 @@ import pytest
 from scipy import sparse
 
 import anndata as ad
-from anndata.compat import CSArray, CSMatrix
 from anndata.tests.helpers import (
     GEN_ADATA_DASK_ARGS,
     GEN_ADATA_NO_XARRAY_ARGS,
@@ -21,6 +20,13 @@ from anndata.tests.helpers import (
 )
 from anndata.utils import asarray
 
+if TYPE_CHECKING:
+    from collections.abc import Callable
+    from pathlib import Path
+    from typing import Literal
+
+    from anndata.compat import DaskArray
+
 subset_func2 = subset_func
 
 
@@ -30,7 +36,7 @@ subset_func2 = subset_func
 
 
 @pytest.fixture
-def adata():
+def adata() -> ad.AnnData:
     X_list = [
         [1, 2, 3],
         [4, 5, 6],
@@ -62,22 +68,26 @@ def adata():
     params=[sparse.csr_matrix, sparse.csc_matrix, np.array, as_dense_dask_array],
     ids=["scipy-csr", "scipy-csc", "np-array", "dask_array"],
 )
-def mtx_format(request):
+def mtx_format(
+    request,
+) -> Callable[
+    [np.ndarray], DaskArray | np.ndarray | sparse.csr_array | sparse.csr_matrix
+]:
     return request.param
 
 
 @pytest.fixture(params=[sparse.csr_matrix, sparse.csc_matrix])
-def sparse_format(request):
+def sparse_format(request) -> type[sparse.csr_matrix | sparse.csc_matrix]:
     return request.param
 
 
 @pytest.fixture(params=["r+", "r", False])
-def backed_mode(request):
+def backed_mode(request) -> Literal["r+", "r", False]:
     return request.param
 
 
 @pytest.fixture(params=(("X",), ()))
-def as_dense(request):
+def as_dense(request) -> tuple[str] | tuple:
     return request.param
 
 
@@ -90,10 +100,17 @@ def as_dense(request):
 @pytest.mark.filterwarnings("ignore:`product` is deprecated as of NumPy 1.25.0")
 # TODO: Check to make sure obs, obsm, layers, ... are written and read correctly as well
 @pytest.mark.filterwarnings("error")
-def test_read_write_X(tmp_path, mtx_format, backed_mode, as_dense):
-    base_pth = Path(tmp_path)
-    orig_pth = base_pth / "orig.h5ad"
-    backed_pth = base_pth / "backed.h5ad"
+def test_read_write_X(
+    tmp_path: Path,
+    mtx_format: Callable[
+        [np.ndarray], DaskArray | np.ndarray | sparse.csr_array | sparse.csr_matrix
+    ],
+    backed_mode: Literal["r+", "r", False],
+    *,
+    as_dense: tuple[str] | tuple,
+):
+    orig_pth = tmp_path / "orig.h5ad"
+    backed_pth = tmp_path / "backed.h5ad"
 
     orig = ad.AnnData(mtx_format(asarray(sparse.random(10, 10, format="csr"))))
     orig.write(orig_pth)
@@ -104,6 +121,21 @@ def test_read_write_X(tmp_path, mtx_format, backed_mode, as_dense):
 
     from_backed = ad.read_h5ad(backed_pth)
     assert np.all(asarray(orig.X) == asarray(from_backed.X))
+
+
+def test_backed_view(tmp_path: Path, backed_mode: Literal["r+", "r", False]):
+    orig_pth = tmp_path / "orig.h5ad"
+
+    orig = ad.AnnData(sparse.random(100, 10, format="csr"))
+    orig.write(orig_pth)
+
+    adata = ad.read_h5ad(orig_pth, backed=backed_mode)
+
+    for i in range(0, adata.shape[0], 10):
+        chunk_path = tmp_path / f"chunk_{i}.h5ad"
+        adata[i : i + 5].write_h5ad(tmp_path / f"chunk_{i}.h5ad")
+        chunk = adata[i : i + 5]
+        assert_equal(chunk, ad.read_h5ad(chunk_path))
 
 
 # this is very similar to the views test
@@ -121,8 +153,8 @@ def test_backing(adata: ad.AnnData, tmp_path: Path, backing_h5ad: Path) -> None:
     # know that the file is open again....
     assert adata.file.is_open
 
-    adata[:2, 0].X = [0, 0]
-    assert adata[:, 0].X.tolist() == np.reshape([0, 0, 7], (3, 1)).tolist()
+    adata[:2, 0].X = np.array([[0], [0]])
+    assert adata[:, 0].X.tolist() == np.reshape([1, 4, 7], (3, 1)).tolist()
 
     adata_subset = adata[:2, [0, 1]]
     assert adata_subset.is_view
@@ -157,7 +189,7 @@ def test_backing(adata: ad.AnnData, tmp_path: Path, backing_h5ad: Path) -> None:
     adata_subset.write()
 
 
-def test_backing_copy(adata, tmp_path, backing_h5ad):
+def test_backing_copy(adata, tmp_path: Path, backing_h5ad: Path):
     adata.filename = backing_h5ad
     adata.write()
 
@@ -171,7 +203,7 @@ def test_backing_copy(adata, tmp_path, backing_h5ad):
 
 
 # TODO: Also test updating the backing file inplace
-def test_backed_raw(tmp_path):
+def test_backed_raw(tmp_path: Path):
     backed_pth = tmp_path / "backed.h5ad"
     final_pth = tmp_path / "final.h5ad"
     mem_adata = gen_adata((10, 10), **GEN_ADATA_DASK_ARGS)
@@ -194,21 +226,20 @@ def test_backed_raw(tmp_path):
         pytest.param(sparse.csr_array, id="csr_array"),
     ],
 )
-def test_backed_raw_subset(tmp_path, array_type, subset_func, subset_func2):
+def test_backed_raw_subset(
+    tmp_path: Path,
+    array_type: Callable[
+        [np.ndarray], np.ndarray | sparse.csr_array | sparse.csr_matrix
+    ],
+    subset_func: Callable[[ad.AnnData], ad.AnnData],
+    subset_func2: Callable[[ad.AnnData], ad.AnnData],
+):
     backed_pth = tmp_path / "backed.h5ad"
     final_pth = tmp_path / "final.h5ad"
     mem_adata = gen_adata((10, 10), X_type=array_type, **GEN_ADATA_NO_XARRAY_ARGS)
     mem_adata.raw = mem_adata
     obs_idx = subset_func(mem_adata.obs_names)
     var_idx = subset_func2(mem_adata.var_names)
-    if (
-        array_type is asarray
-        and isinstance(obs_idx, list | np.ndarray | CSMatrix | CSArray)
-        and isinstance(var_idx, list | np.ndarray | CSMatrix | CSArray)
-    ):
-        pytest.xfail(
-            "Fancy indexing does not work with multiple arrays on a h5py.Dataset"
-        )
     mem_adata.write(backed_pth)
 
     ### Backed view has same values as in memory view ###
@@ -240,7 +271,10 @@ def test_backed_raw_subset(tmp_path, array_type, subset_func, subset_func2):
         pytest.param(as_dense_dask_array, id="dask_array"),
     ],
 )
-def test_to_memory_full(tmp_path, array_type):
+def test_to_memory_full(
+    tmp_path: Path,
+    array_type: Callable[[np.ndarray], np.ndarray | DaskArray | sparse.csr_matrix],
+):
     backed_pth = tmp_path / "backed.h5ad"
     mem_adata = gen_adata((15, 10), X_type=array_type, **GEN_ADATA_DASK_ARGS)
     mem_adata.raw = gen_adata((15, 12), X_type=array_type, **GEN_ADATA_DASK_ARGS)
@@ -255,7 +289,7 @@ def test_to_memory_full(tmp_path, array_type):
     assert_equal(mem_adata, backed_adata.to_memory())
 
 
-def test_double_index(adata, backing_h5ad):
+def test_double_index(adata: ad.AnnData, backing_h5ad: Path):
     adata.filename = backing_h5ad
     with pytest.raises(ValueError, match=r"cannot make a view of a view"):
         # no view of view of backed object currently
@@ -265,7 +299,7 @@ def test_double_index(adata, backing_h5ad):
     adata.write()
 
 
-def test_return_to_memory_mode(adata, backing_h5ad):
+def test_return_to_memory_mode(adata: ad.AnnData, backing_h5ad: Path):
     bdata = adata.copy()
     adata.filename = backing_h5ad
     assert adata.isbacked
@@ -283,7 +317,7 @@ def test_return_to_memory_mode(adata, backing_h5ad):
     bdata.filename = None
 
 
-def test_backed_modification(adata, backing_h5ad):
+def test_backed_modification(adata: ad.AnnData, backing_h5ad: Path):
     adata.X[:, 1] = 0  # Make it a little sparse
     adata.X = sparse.csr_matrix(adata.X)
     assert not adata.isbacked
@@ -305,9 +339,14 @@ def test_backed_modification(adata, backing_h5ad):
     assert np.all(adata.X[2, :] == np.array([7, 13, 9]))
 
 
-def test_backed_modification_sparse(adata, backing_h5ad, sparse_format):
+def test_backed_modification_sparse(
+    adata: ad.AnnData,
+    backing_h5ad: Path,
+    sparse_format: type[sparse.csr_matrix | sparse.csc_matrix],
+):
     adata.X[:, 1] = 0  # Make it a little sparse
     adata.X = sparse_format(adata.X)
+    orig = adata.X.copy()
     assert not adata.isbacked
 
     adata.write(backing_h5ad)
@@ -316,22 +355,9 @@ def test_backed_modification_sparse(adata, backing_h5ad, sparse_format):
     assert adata.filename == backing_h5ad
     assert adata.isbacked
 
-    pat = r"__setitem__ for backed sparse will be removed"
-    with pytest.warns(FutureWarning, match=pat):
-        adata.X[0, [0, 2]] = 10
-    with pytest.warns(FutureWarning, match=pat):
-        adata.X[1, [0, 2]] = [11, 12]
-    with (
-        pytest.warns(FutureWarning, match=pat),
-        pytest.raises(ValueError, match=r"cannot change the sparsity structure"),
-    ):
-        adata.X[2, 1] = 13
-
-    assert adata.isbacked
-
-    assert np.all(adata.X[0, :] == np.array([10, 0, 10]))
-    assert np.all(adata.X[1, :] == np.array([11, 0, 12]))
-    assert np.all(adata.X[2, :] == np.array([7, 0, 9]))
+    # Does not modify backed store
+    adata[0, [0, 2]].X = np.array([[10, 10]])
+    np.testing.assert_equal(orig.toarray(), adata.X[...].toarray())
 
 
 # TODO: Work around h5py not supporting this
@@ -355,3 +381,105 @@ def test_backed_modification_sparse(adata, backing_h5ad, sparse_format):
 #     backed_view = backed_adata[[1,2], :]
 #     backed_view.X = 0
 #     assert np.all(backed_adata.X[[1,2], :] == 0)
+
+
+@pytest.mark.parametrize(
+    ("obs_idx", "var_idx"),
+    [
+        pytest.param(np.array([0, 1, 2]), np.array([1, 2]), id="no_dupes"),
+        pytest.param(np.array([0, 1, 0, 2]), slice(None), id="1d_dupes"),
+        pytest.param(np.array([0, 1, 0, 2]), np.array([1, 2, 1]), id="2d_dupes"),
+    ],
+)
+def test_backed_duplicate_indices(tmp_path, obs_idx, var_idx):
+    """Test that backed HDF5 datasets handle duplicate indices correctly."""
+    backed_pth = tmp_path / "backed.h5ad"
+
+    # Create test data
+    mem_adata = gen_adata((6, 4), X_type=asarray, **GEN_ADATA_NO_XARRAY_ARGS)
+    mem_adata.write(backed_pth)
+
+    # Load backed data
+    backed_adata = ad.read_h5ad(backed_pth, backed="r")
+
+    # Test the indexing
+    mem_result_multi = mem_adata[obs_idx, var_idx]
+    backed_result_multi = backed_adata[obs_idx, var_idx]
+    assert_equal(mem_result_multi, backed_result_multi)
+
+
+@pytest.fixture
+def h5py_test_data(tmp_path):
+    """Create test HDF5 file with dataset for _safe_fancy_index_h5py tests."""
+    import h5py
+
+    h5_path = tmp_path / "test_dataset.h5"
+    test_data = np.arange(24).reshape(6, 4)  # 6x4 matrix
+
+    with h5py.File(h5_path, "w") as f:
+        f.create_dataset("test", data=test_data)
+
+    return h5_path, test_data
+
+
+@pytest.mark.parametrize(
+    ("indices", "description"),
+    [
+        pytest.param((np.array([0, 1, 0, 2]),), "single_dimension_with_duplicates"),
+        pytest.param(
+            (np.array([0, 1, 2]), np.array([1, 2])), "multi_dimensional_no_duplicates"
+        ),
+        pytest.param(
+            (np.array([0, 1, 0, 2]), np.array([1, 2])),
+            "multi_dimensional_duplicates_first_dim",
+        ),
+        pytest.param(
+            (np.array([0, 1, 2]), np.array([1, 2, 1])),
+            "multi_dimensional_duplicates_second_dim",
+        ),
+        pytest.param(
+            (np.array([0, 1, 0]), np.array([1, 2, 1])),
+            "multi_dimensional_duplicates_both_dims",
+        ),
+        pytest.param(
+            (np.array([True, False, True, False, False, True]),), "boolean_arrays"
+        ),
+        pytest.param((np.array([0, 1, 0]), slice(1, 3)), "mixed_indexing_with_slices"),
+        pytest.param(
+            (np.array([0, 1, 0]), [1, 2]), "mixed_indexing_with_slices_and_lists"
+        ),
+        pytest.param((np.array([3, 1, 3, 0, 1]),), "unsorted_indices_with_duplicates"),
+    ],
+)
+def test_safe_fancy_index_h5py_function(h5py_test_data, indices, description):
+    """Test the _safe_fancy_index_h5py function directly with various indexing patterns."""
+    import h5py
+
+    from anndata._core.index import _safe_fancy_index_h5py
+
+    h5_path, test_data = h5py_test_data
+
+    with h5py.File(h5_path, "r") as f:
+        dataset = f["test"]
+
+        # Get result from the function
+        result = _safe_fancy_index_h5py(dataset, indices)
+
+    # Calculate expected result using NumPy
+    if isinstance(indices, tuple) and len(indices) > 1:
+        # Multi-dimensional case - use np.ix_ for fancy indexing
+        if isinstance(indices[1], slice):
+            # Handle mixed case with slice
+            expected = test_data[
+                np.ix_(indices[0], np.arange(indices[1].start, indices[1].stop))
+            ]
+        else:
+            expected = test_data[np.ix_(*indices)]
+    else:
+        # Single dimensional case
+        expected = test_data[indices]
+
+    # Assert arrays are equal
+    np.testing.assert_array_equal(
+        result, expected, err_msg=f"Failed for test case: {description}"
+    )
