@@ -2,7 +2,7 @@ from __future__ import annotations
 
 from functools import partial
 from itertools import product
-from typing import TYPE_CHECKING, Literal, get_args
+from typing import TYPE_CHECKING, Literal
 
 import h5py
 import numpy as np
@@ -15,10 +15,11 @@ from anndata._core.anndata import AnnData
 from anndata._core.sparse_dataset import sparse_dataset
 from anndata._io.specs.registry import read_elem_lazy
 from anndata._io.zarr import open_write_group
-from anndata.compat import CSArray, CSMatrix, DaskArray, ZarrGroup, is_zarr_v2
+from anndata.compat import CSArray, CSMatrix, DaskArray, ZarrGroup
 from anndata.experimental import read_dispatched
 from anndata.tests import helpers as test_helpers
 from anndata.tests.helpers import AccessTrackingStore, assert_equal, subset_func
+from anndata.utils import get_literal_members
 
 if TYPE_CHECKING:
     from collections.abc import Callable, Generator, Sequence
@@ -31,7 +32,7 @@ if TYPE_CHECKING:
 
     from anndata.abc import CSCDataset, CSRDataset
 
-    Idx = slice | int | NDArray[np.integer] | NDArray[np.bool_]
+    type Idx = slice | int | NDArray[np.integer] | NDArray[np.bool_]
 
 
 subset_func2 = subset_func
@@ -83,12 +84,9 @@ def ondisk_equivalent_adata(
             # Read with handling for backwards compat
             def callback(func, elem_name, elem, iospec):
                 if iospec.encoding_type == "anndata" or elem_name.endswith("/"):
-                    return AnnData(
-                        **{
-                            k: read_dispatched(v, callback)
-                            for k, v in dict(elem).items()
-                        }
-                    )
+                    return AnnData(**{
+                        k: read_dispatched(v, callback) for k, v in dict(elem).items()
+                    })
                 if iospec.encoding_type in {"csc_matrix", "csr_matrix"}:
                     return sparse_dataset(elem)
                 return func(elem)
@@ -225,7 +223,9 @@ def test_consecutive_bool(
     # `_normalize_indices`, which is used by `AnnData`, converts bools to ints with `np.where`
     from anndata._core import sparse_dataset
 
-    spy = mocker.spy(sparse_dataset, "get_compressed_vectors_for_slices")
+    spy = mocker.spy(
+        sparse_dataset.BackedSparseMatrix, "get_compressed_vectors_for_slices"
+    )
     assert_equal(csr_disk.X[mask, :], csr_disk.X[np.where(mask)])
     if should_trigger_optimization is not None:
         assert (
@@ -387,7 +387,7 @@ def test_lazy_array_cache(
     a = sparse_format(sparse.random(10, 10))
     f = open_write_group(path, mode="a")
     ad.io.write_elem(f, "X", a)
-    store = AccessTrackingStore(path)
+    store = AccessTrackingStore(path, read_only=True)
     for elem in elems:
         store.initialize_key_trackers([f"X/{elem}"])
     f = zarr.open_group(store, mode="r")
@@ -396,18 +396,11 @@ def test_lazy_array_cache(
     a_disk[3:5]
     a_disk[6:7]
     a_disk[8:9]
-    # One hit for .zarray in zarr v2 and three for metadata in zarr v3:
+    # Three hits for metadata in zarr v3:
     # see https://github.com/zarr-developers/zarr-python/discussions/2760 for more info on the difference.
     # Then there is actual data access, 1 more when cached, 4 more otherwise.
-    match should_cache_indptr, is_zarr_v2():
-        case True, True:
-            assert store.get_access_count("X/indptr") == 2
-        case False, True:
-            assert store.get_access_count("X/indptr") == 5
-        case True, False:
-            assert store.get_access_count("X/indptr") == 4
-        case False, False:
-            assert store.get_access_count("X/indptr") == 7
+    c_expected = 4 if should_cache_indptr else 7
+    assert store.get_access_count("X/indptr") == c_expected
     for elem_not_indptr in elems - {"indptr"}:
         assert (
             sum(
@@ -418,7 +411,7 @@ def test_lazy_array_cache(
         )
 
 
-Kind = Literal["slice", "int", "array", "mask"]
+type Kind = Literal["slice", "int", "array", "mask"]
 
 
 def mk_idx_kind(idx: Sequence[int], *, kind: Kind, l: int) -> Idx | None:
@@ -450,7 +443,9 @@ def width_idx_kinds(
     *idxs: tuple[Sequence[int], Idx, Sequence[str]], l: int
 ) -> Generator[ParameterSet, None, None]:
     """Convert major (first) index into various identical kinds of indexing."""
-    for (idx_maj_raw, idx_min, exp), maj_kind in product(idxs, get_args(Kind)):
+    for (idx_maj_raw, idx_min, exp), maj_kind in product(
+        idxs, get_literal_members(Kind)
+    ):
         if (idx_maj := mk_idx_kind(idx_maj_raw, kind=maj_kind, l=l)) is None:
             continue
         id_ = "-".join(map(idify, [idx_maj_raw, idx_min, maj_kind]))
@@ -515,16 +510,13 @@ def test_data_access(
     data = f["X/data"][...]
     del f["X/data"]
     # chunk one at a time to count properly
-    kwargs = {}
-    if not is_zarr_v2():
-        kwargs["zarr_format"] = f.metadata.zarr_format
     zarr.array(
         data,
         store=path / "X" / "data",
         chunks=(1,),
-        **kwargs,
+        zarr_format=f.metadata.zarr_format,
     )
-    store = AccessTrackingStore(path)
+    store = AccessTrackingStore(path, read_only=True)
     store.initialize_key_trackers(["X/data"])
     f = zarr.open_group(store, mode="r")
     a_disk = AnnData(X=open_func(f["X"]))
@@ -533,7 +525,7 @@ def test_data_access(
         subset.X.compute(scheduler="single-threaded")
     # zarr v2 fetches all and not just metadata for that node in 3.X.X python package
     # TODO: https://github.com/zarr-developers/zarr-python/discussions/2760
-    if ad.settings.zarr_write_format == 2 and not is_zarr_v2():
+    if ad.settings.zarr_write_format == 2:
         exp = [*exp, "X/data/.zgroup", "X/data/.zattrs"]
 
     assert store.get_access_count("X/data") == len(exp), store.get_accessed_keys(
@@ -599,7 +591,7 @@ def test_wrong_formats(tmp_path: Path):
         disk_mtx.append(sparse.random(100, 100, format="coo"))
     with pytest.raises(NotImplementedError):
         disk_mtx.append(np.random.random((100, 100)))
-    if isinstance(f, ZarrGroup) and not is_zarr_v2():
+    if isinstance(f, ZarrGroup):
         data = np.random.random((100, 100))
         disk_dense = f.create_array("dense", shape=(100, 100), dtype=data.dtype)
         disk_dense[...] = data
@@ -627,6 +619,23 @@ def test_anndata_sparse_compat(tmp_path: Path, diskfmt: Literal["h5ad", "zarr"])
     assert_equal(adata.X, base)
 
 
+def test_write(tmp_path: Path, diskfmt: Literal["h5ad", "zarr"]):
+    base = sparse.random(10, 10, format="csr")
+
+    f = (
+        open_write_group(tmp_path / f"parent_store.{diskfmt}", mode="a")
+        if diskfmt == "zarr"
+        else h5py.File(tmp_path / f"parent_store.{diskfmt}", "a")
+    )
+
+    ad.io.write_elem(f, "a_sparse_matrix", base)
+    adata = ad.AnnData(sparse_dataset(f["a_sparse_matrix"]))
+    ad.io.write_elem(f, "adata", adata)
+    adata_roundtripped = ad.io.read_elem(f["adata"])
+    assert_equal(adata_roundtripped.X, base)
+
+
+@pytest.mark.filterwarnings(r"ignore:.*array concat.*empty entries:FutureWarning")
 def test_backed_sizeof(
     ondisk_equivalent_adata: tuple[AnnData, AnnData, AnnData, AnnData],
 ):
