@@ -3,21 +3,27 @@ from __future__ import annotations
 from collections.abc import Mapping
 from typing import TYPE_CHECKING
 
+import awkward as ak
+import h5py
 import numpy as np
 import pandas as pd
 import pytest
 from scipy import sparse
 
-from anndata import AnnData, concat
+import anndata as ad
+from anndata import AnnData, concat, settings
+from anndata._core import merge
 from anndata._core.merge import _resolve_axis
 from anndata.experimental.merge import as_group, concat_on_disk
 from anndata.io import read_elem, write_elem
-from anndata.tests.helpers import assert_equal, gen_adata
+from anndata.tests.helpers import assert_equal, check_all_sharded, gen_adata
 from anndata.utils import asarray
 
 if TYPE_CHECKING:
     from pathlib import Path
     from typing import Literal
+
+    from anndata._types import Join_T
 
 
 GEN_ADATA_OOC_CONCAT_ARGS = dict(
@@ -29,6 +35,11 @@ GEN_ADATA_OOC_CONCAT_ARGS = dict(
     varm_types=(sparse.csr_matrix, np.ndarray, pd.DataFrame),
     layers_types=(sparse.csr_matrix, np.ndarray, pd.DataFrame),
 )
+
+
+@pytest.fixture(params=list(merge.MERGE_STRATEGIES.keys()))
+def merge_strategy(request):
+    return request.param
 
 
 @pytest.fixture(params=[0, 1])
@@ -69,13 +80,15 @@ def _adatas_to_paths(adatas, tmp_path, file_format):
         paths = {}
         for k, v in adatas.items():
             p = tmp_path / (f"{k}." + file_format)
-            write_elem(as_group(p, mode="a"), "", v)
+            with as_group(p, mode="a") as f:
+                write_elem(f, "", v)
             paths[k] = p
     else:
         paths = []
         for i, a in enumerate(adatas):
             p = tmp_path / (f"{i}." + file_format)
-            write_elem(as_group(p, mode="a"), "", a)
+            with as_group(p, mode="a") as f:
+                write_elem(f, "", a)
             paths += [p]
     return paths
 
@@ -86,17 +99,19 @@ def assert_eq_concat_on_disk(
     file_format: Literal["zarr", "h5ad"],
     max_loaded_elems: int | None = None,
     *args,
+    merge_strategy: merge.StrategiesLiteral | None = None,
     **kwargs,
 ):
     # create one from the concat function
-    res1 = concat(adatas, *args, **kwargs)
+    res1 = concat(adatas, *args, merge=merge_strategy, **kwargs)
     # create one from the on disk concat function
     paths = _adatas_to_paths(adatas, tmp_path, file_format)
     out_name = tmp_path / f"out.{file_format}"
     if max_loaded_elems is not None:
         kwargs["max_loaded_elems"] = max_loaded_elems
-    concat_on_disk(paths, out_name, *args, **kwargs)
-    res2 = read_elem(as_group(out_name, mode="r"))
+    concat_on_disk(paths, out_name, *args, merge=merge_strategy, **kwargs)
+    with as_group(out_name, mode="r") as rg:
+        res2 = read_elem(rg)
     assert_equal(res1, res2, exact=False)
 
 
@@ -112,15 +127,17 @@ def get_array_type(array_type, axis):
 
 
 @pytest.mark.parametrize("reindex", [True, False], ids=["reindex", "no_reindex"])
+@pytest.mark.filterwarnings("ignore:Misaligned chunks detected")
 def test_anndatas(
     *,
     axis: Literal[0, 1],
     array_type: Literal["array", "sparse", "sparse_array"],
-    join_type: Literal["inner", "outer"],
+    join_type: Join_T,
     tmp_path: Path,
     max_loaded_elems: int,
     file_format: Literal["zarr", "h5ad"],
     reindex: bool,
+    merge_strategy: merge.StrategiesLiteral,
 ):
     _, off_axis_name = _resolve_axis(1 - axis)
     random_axes = {0, 1} if reindex else {axis}
@@ -159,6 +176,7 @@ def test_anndatas(
         max_loaded_elems,
         axis=axis,
         join=join_type,
+        merge_strategy=merge_strategy,
     )
 
 
@@ -220,7 +238,7 @@ def xxxm_adatas():
             X=sparse.csr_matrix((2, 100)),
             obs=pd.DataFrame(index=gen_index(2)),
             obsm={
-                "sparse": np.arange(8).reshape(2, 4),
+                "sparse": sparse.csr_matrix(np.arange(8).reshape(2, 4)),
                 "dense": np.arange(4, 8).reshape(2, 2),
                 "df": pd.DataFrame(
                     {
@@ -243,14 +261,76 @@ def test_concatenate_xxxm(xxxm_adatas, tmp_path, file_format, join_type):
     assert_eq_concat_on_disk(xxxm_adatas, tmp_path, file_format, join=join_type)
 
 
+def test_concatenate_zarr_v3_shard(xxxm_adatas, tmp_path):
+    import zarr
+
+    with settings.override(auto_shard_zarr_v3=True, zarr_write_format=3):
+        assert_eq_concat_on_disk(xxxm_adatas, tmp_path, file_format="zarr")
+    g = zarr.open(tmp_path)
+    assert g.metadata.zarr_format == 3
+
+    check_all_sharded(g)
+
+
+def test_singleton(xxxm_adatas, tmp_path, file_format):
+    assert_eq_concat_on_disk(xxxm_adatas[:1], tmp_path, file_format=file_format)
+
+
 def test_output_dir_exists(tmp_path):
     in_pth = tmp_path / "in.h5ad"
     out_pth = tmp_path / "does_not_exist" / "out.h5ad"
 
     AnnData(X=np.ones((5, 1))).write_h5ad(in_pth)
 
-    with pytest.raises(FileNotFoundError, match=f"{out_pth}"):
+    with pytest.raises(FileNotFoundError, match=str(out_pth)):
         concat_on_disk([in_pth], out_pth)
+
+
+def test_no_open_h5_file_handles_after_error(tmp_path):
+    in_pth = tmp_path / "in.h5ad"
+    in_pth2 = tmp_path / "in2.h5ad"
+    out_pth = tmp_path / "out.h5ad"
+
+    adata = AnnData(
+        X=np.ones((2, 1)),
+        obsm={
+            "awk": ak.Array([
+                [{"a": 1, "b": "foo"}],
+                [{"a": 2, "b": "bar"}],
+            ])
+        },
+    )
+    adata.write_h5ad(in_pth)
+    adata.write_h5ad(in_pth2)
+
+    # Intentionally write an unsupported array type, which could leave dangling file handles:
+    # https://github.com/scverse/anndata/issues/2198
+    try:
+        concat_on_disk([in_pth, in_pth2], out_pth)
+    except NotImplementedError:
+        for path in [in_pth, in_pth2, out_pth]:
+            # should not error out because there are no file handles open
+            f = h5py.File(path, mode="w")
+            f.close()
+
+
+def test_write_using_groups(tmp_path, file_format):
+    in_pth = tmp_path / f"in.{file_format}"
+    in_pth2 = tmp_path / f"in2.{file_format}"
+    out_pth = tmp_path / f"out.{file_format}"
+
+    adata = AnnData(X=np.ones((2, 1)))
+    getattr(adata, f"write_{file_format}")(in_pth)
+    getattr(adata, f"write_{file_format}")(in_pth2)
+
+    with (
+        as_group(in_pth, mode="r") as f1,
+        as_group(in_pth2, mode="r") as f2,
+        as_group(out_pth, mode="w") as fout,
+    ):
+        concat_on_disk([f1, f2], fout)
+    adata_out = getattr(ad, f"read_{file_format}")(out_pth)
+    assert_equal(adata_out, concat([adata, adata]))
 
 
 def test_failure_w_no_args(tmp_path):
