@@ -26,7 +26,12 @@ def attr_name(request):
     return request.param
 
 
-@pytest.fixture(params=[True, False])
+@pytest.fixture(
+    params=[
+        pytest.param(True, id="give_chunks"),
+        pytest.param(False, id="no_give_chunks"),
+    ]
+)
 def give_chunks(request):
     return request.param
 
@@ -69,8 +74,9 @@ def _alloc_cache():
 # if we put a 2 factor on 2**19
 # the results seems more accurate with the experimental results
 # For example from dask.random we allocate 1mb
+# See https://github.com/scverse/anndata/issues/2301 for why this limit is 3.0 MB
 @pytest.mark.usefixtures("_alloc_cache")
-@pytest.mark.limit_memory("1.5 MB")
+@pytest.mark.limit_memory("3.0 MB")
 def test_size_of_view(mapping_name, give_chunks):
     import dask.array as da
 
