@@ -10,7 +10,7 @@ import pytest
 import zarr
 from scipy.sparse import csr_matrix
 
-from anndata import AnnData
+from anndata import AnnData, settings
 from anndata._io.specs.registry import read_elem_partial
 from anndata.io import read_elem, write_h5ad, write_zarr
 
@@ -43,17 +43,20 @@ def test_read_partial_X(tmp_path, typ, diskfmt):
 
 @pytest.mark.skipif(not find_spec("scanpy"), reason="Scanpy is not installed")
 def test_read_partial_adata(tmp_path, diskfmt):
+    import scanpy as sc
+
     with warnings.catch_warnings():
         warnings.filterwarnings(
-            "ignore", message=r"Importing read_.* from `anndata` is deprecated"
+            "ignore", r"Moving element.*uns.*to.*obsp", FutureWarning
         )
-        import scanpy as sc
-
-    adata = sc.datasets.pbmc68k_reduced()
+        adata = sc.datasets.pbmc68k_reduced()
 
     path = Path(tmp_path) / ("test_rp." + diskfmt)
 
-    WRITER[diskfmt](path, adata)
+    # we’re not adding things to read_partial anymore, so it can only read non-nullable strings.
+    # therefore force writing old format here
+    with settings.override(allow_write_nullable_strings=False):
+        WRITER[diskfmt](path, adata)
 
     storage = READER[diskfmt](path, mode="r")
 
