@@ -10,12 +10,13 @@ import pytest
 import zarr
 
 import anndata as ad
-from anndata._io.specs.registry import IORegistryError
+from anndata._io.specs.registry import IORegistryError, to_writeable
 from anndata._io.utils import report_read_key_on_error
 from anndata.compat import _clean_uns
+from anndata.tests.helpers import jnp
 
 if TYPE_CHECKING:
-    from collections.abc import Callable
+    from collections.abc import Callable, Iterator
     from pathlib import Path
 
 
@@ -110,3 +111,41 @@ def test_only_child_key_reported_on_failure(tmp_path, group_fn):
 
     with pytest.raises(IORegistryError, match=pattern):
         ad.io.read_elem(group)
+
+
+@pytest.mark.parametrize("to", [np.array, pd.Series, lambda x: x])
+def test_to_writeable_passthrough(to: Callable[[list], np.ndarray | pd.Series | list]):
+    x = to([1, 2, 3])
+    result = to_writeable(x)
+    assert result is x
+
+
+def get_gpu_devices() -> Iterator:
+    return (
+        d
+        for d in jnp.__array_namespace_info__().devices()
+        if d is not None and d.device_kind == "GPU"
+    )
+
+
+skip_if_no_jax_gpu = pytest.mark.skipif(
+    jnp is None or not any(get_gpu_devices()),
+    reason="No GPU available",
+)
+
+
+@skip_if_no_jax_gpu
+@pytest.mark.array_api
+def test_to_writeable_jax_gpu_array() -> None:
+    x = jnp.asarray([1.0, 2.0], device=get_gpu_devices()[0])
+    result = to_writeable(x)
+    assert isinstance(result, np.ndarray)
+    np.testing.assert_array_equal(result, np.array([1.0, 2.0]))
+
+
+@pytest.mark.array_api
+def test_to_writeable_does_not_recurse() -> None:
+    x = {"a": jnp.array([1.0, 2.0])}
+    result = to_writeable(x)
+    # since dict is not supported, it should return unchanged
+    assert result is x
