@@ -15,7 +15,6 @@ import pandas as pd
 import pytest
 import zarr
 import zarr.convenience
-from numba.core.errors import NumbaDeprecationWarning
 from scipy.sparse import csc_array, csc_matrix, csr_array, csr_matrix
 
 import anndata as ad
@@ -28,21 +27,28 @@ from anndata.compat import (
     ZarrArray,
     ZarrGroup,
     _read_attr,
-    is_zarr_v2,
 )
 from anndata.tests.helpers import (
     GEN_ADATA_NO_XARRAY_ARGS,
     as_dense_dask_array,
     assert_equal,
     gen_adata,
+    jnp,
+    jnp_array_or_idempotent,
 )
 
 if TYPE_CHECKING:
-    from collections.abc import Generator
+    from collections.abc import Callable, Generator
     from typing import Literal
 
 HERE = Path(__file__).parent
-
+ARRAY_TYPES = [
+    pytest.param(np.array, id="np.array"),
+    pytest.param(as_dense_dask_array, id="dask_dense"),
+    pytest.param(csr_matrix, id="csr_matrix"),
+    pytest.param(csr_array, id="csr_array"),
+    pytest.param(jnp_array_or_idempotent, id="jax.array", marks=pytest.mark.array_api),
+]
 
 # ------------------------------------------------------------------------------
 # Some test data
@@ -120,7 +126,7 @@ def dtype(request):
 # ------------------------------------------------------------------------------
 
 
-@pytest.mark.parametrize("typ", [np.array, csr_matrix, csr_array, as_dense_dask_array])
+@pytest.mark.parametrize("typ", ARRAY_TYPES)
 def test_readwrite_roundtrip(typ, tmp_path, diskfmt, diskfmt2):
     pth1 = tmp_path / f"first.{diskfmt}"
     write1 = lambda x: getattr(x, f"write_{diskfmt}")(pth1)
@@ -160,8 +166,14 @@ def test_readwrite_roundtrip_async(tmp_path):
 
 
 @pytest.mark.parametrize("storage", ["h5ad", "zarr"])
-@pytest.mark.parametrize("typ", [np.array, csr_matrix, csr_array, as_dense_dask_array])
-def test_readwrite_kitchensink(tmp_path, storage, typ, backing_h5ad, dataset_kwargs):
+@pytest.mark.parametrize("typ", ARRAY_TYPES)
+def test_readwrite_kitchensink(
+    tmp_path: Path,
+    storage: Literal["h5ad", "zarr"],
+    typ,
+    backing_h5ad: Path,
+    dataset_kwargs,
+) -> None:
     X = typ(X_list)
     adata_src = ad.AnnData(X, obs=obs_dict, var=var_dict, uns=uns_dict)
     assert not isinstance(adata_src.obs["oanno1"].dtype, pd.CategoricalDtype)
@@ -186,18 +198,20 @@ def test_readwrite_kitchensink(tmp_path, storage, typ, backing_h5ad, dataset_kwa
     assert_equal(adata.var.index, adata_src.var.index)
     assert adata.var.index.dtype == adata_src.var.index.dtype
 
+    used_jax = jnp is not None and isinstance(adata_src.X, jnp.ndarray)
     # Dev. Note:
     # either load as same type or load the convert DaskArray to array
     # since we tested if assigned types and loaded types are DaskArray
     # this would also work if they work
     if isinstance(adata_src.raw.X, CSArray):
         assert isinstance(adata.raw.X, CSMatrix)
-    else:
+    elif not used_jax:
         assert isinstance(adata_src.raw.X, type(adata.raw.X) | DaskArray)
     assert isinstance(
         adata_src.uns["uns4"]["c"], type(adata.uns["uns4"]["c"]) | DaskArray
     )
-    assert isinstance(adata_src.varm, type(adata.varm) | DaskArray)
+    if not used_jax:
+        assert isinstance(adata_src.varm, type(adata.varm) | DaskArray)
 
     assert_equal(adata.raw.X, adata_src.raw.X)
     pd.testing.assert_frame_equal(adata.raw.var, adata_src.raw.var)
@@ -206,13 +220,15 @@ def test_readwrite_kitchensink(tmp_path, storage, typ, backing_h5ad, dataset_kwa
     assert_equal(adata, adata_src)
 
 
-@pytest.mark.parametrize("typ", [np.array, csr_matrix, csr_array, as_dense_dask_array])
-def test_readwrite_maintain_X_dtype(typ, backing_h5ad):
+@pytest.mark.parametrize("typ", ARRAY_TYPES)
+def test_readwrite_maintain_X_dtype(typ, backing_h5ad: Path) -> None:
     X = typ(X_list).astype("int8")
     adata_src = ad.AnnData(X)
     adata_src.write(backing_h5ad)
 
     adata = ad.read_h5ad(backing_h5ad)
+    if jnp is not None and isinstance(adata_src.X, jnp.ndarray):
+        adata_src.X = np.from_dlpack(adata_src.X)
     assert adata.X.dtype == adata_src.X.dtype
 
 
@@ -250,8 +266,8 @@ def test_readwrite_h5ad_one_dimension(typ, backing_h5ad):
     assert_equal(adata, adata_one)
 
 
-@pytest.mark.parametrize("typ", [np.array, csr_matrix, csr_array, as_dense_dask_array])
-def test_readwrite_backed(typ, backing_h5ad):
+@pytest.mark.parametrize("typ", ARRAY_TYPES)
+def test_readwrite_backed(typ, backing_h5ad: Path) -> None:
     X = typ(X_list)
     adata_src = ad.AnnData(X, obs=obs_dict, var=var_dict, uns=uns_dict)
     adata_src.filename = backing_h5ad  # change to backed mode
@@ -266,9 +282,19 @@ def test_readwrite_backed(typ, backing_h5ad):
 
 
 @pytest.mark.parametrize(
-    "typ", [np.array, csr_matrix, csc_matrix, csr_array, csc_array]
+    "typ",
+    [
+        pytest.param(np.array, id="np.array"),
+        pytest.param(csr_matrix, id="csr_matrix"),
+        pytest.param(csc_matrix, id="csc_matrix"),
+        pytest.param(csr_array, id="csr_array"),
+        pytest.param(csc_array, id="csc_array"),
+        pytest.param(
+            jnp_array_or_idempotent, id="jax.array", marks=pytest.mark.array_api
+        ),
+    ],
 )
-def test_readwrite_equivalent_h5ad_zarr(tmp_path, typ):
+def test_readwrite_equivalent_h5ad_zarr(tmp_path: Path, typ) -> None:
     h5ad_pth = tmp_path / "adata.h5ad"
     zarr_pth = tmp_path / "adata.zarr"
 
@@ -308,14 +334,15 @@ def test_read_full_io_error(tmp_path, name, read, write):
     path = tmp_path / name
     write(adata, path)
     with store_context(path) as store:
-        if not is_zarr_v2() and isinstance(store, ZarrGroup):
+        if isinstance(store, ZarrGroup):
             # see https://github.com/zarr-developers/zarr-python/issues/2716 for the issue
             # with re-opening without syncing attributes explicitly
             # TODO: Having to fully specify attributes to not override fixed in zarr v3.0.5
             # See https://github.com/zarr-developers/zarr-python/pull/2870
-            store["obs"].update_attributes(
-                {**dict(store["obs"].attrs), "encoding-type": "invalid"}
-            )
+            store["obs"].update_attributes({
+                **dict(store["obs"].attrs),
+                "encoding-type": "invalid",
+            })
             zarr.consolidate_metadata(store.store)
         else:
             store["obs"].attrs["encoding-type"] = "invalid"
@@ -377,56 +404,49 @@ def test_hdf5_compression_opts(tmp_path, compression, compression_opts):
 
 
 @pytest.mark.parametrize("zarr_write_format", [2, 3])
-def test_zarr_compression(tmp_path, zarr_write_format):
-    if zarr_write_format == 3 and is_zarr_v2():
-        pytest.xfail("Cannot write zarr v3 format with v2 package")
+@pytest.mark.parametrize(
+    "use_compression", [True, False], ids=["compressed", "uncompressed"]
+)
+def test_zarr_compression(
+    tmp_path: Path, zarr_write_format: Literal[2, 3], *, use_compression: bool
+):
     ad.settings.zarr_write_format = zarr_write_format
     pth = str(Path(tmp_path) / "adata.zarr")
     adata = gen_adata((10, 8), **GEN_ADATA_NO_XARRAY_ARGS)
-    if zarr_write_format == 2 or is_zarr_v2():
+    if not use_compression:
+        compressor = None
+    elif zarr_write_format == 2:
         from numcodecs import Blosc
 
         compressor = Blosc(cname="zstd", clevel=3, shuffle=Blosc.BITSHUFFLE)
     else:
-        from zarr.codecs import BloscCodec
+        from zarr.codecs import ZstdCodec
 
-        # Typesize is forced to be 1 so that the codecs always match on the roundtrip.
-        # Otherwise this value would vary depending on the datatype.
-        # See github.com/zarr-developers/numcodecs/pull/713 for a related issue/explanation.
-        # In practice, you would never want to set this parameter.
-        compressor = BloscCodec(
-            cname="zstd", clevel=3, shuffle="bitshuffle", typesize=1
-        )
-    not_compressed = []
+        # Don't use Blosc since it's defaults can change:
+        # https://github.com/zarr-developers/zarr-python/pull/3545
+        compressor = ZstdCodec(level=3, checksum=True)
+    wrongly_compressed = []
 
     ad.io.write_zarr(pth, adata, compressor=compressor)
 
     def check_compressed(value, key):
         if not isinstance(value, ZarrArray) or value.shape == ():
             return None
-        (read_compressor,) = value.compressors
+        (read_compressor,) = value.compressors or [None]
         if zarr_write_format == 2:
             if read_compressor != compressor:
-                not_compressed.append(key)
+                wrongly_compressed.append(key)
             return None
-        if read_compressor.to_dict() != compressor.to_dict():
-            not_compressed.append(key)
+        if (compressor is None and read_compressor is not None) or (
+            None not in {compressor, read_compressor}
+            and read_compressor.to_dict() != compressor.to_dict()
+        ):
+            wrongly_compressed.append(key)
 
-    if is_zarr_v2():
-        with zarr.open(pth, "r") as f:
-            f.visititems(check_compressed)
-    else:
-        f = zarr.open(pth, mode="r")
-        for key, value in f.members(max_depth=None):
-            check_compressed(value, key)
-
-    if not_compressed:
-        sep = "\n\t"
-        msg = (
-            f"These elements were not compressed correctly:{sep}"
-            f"{sep.join(not_compressed)}"
-        )
-        raise AssertionError(msg)
+    f = zarr.open(pth, mode="r")
+    for key, value in f.members(max_depth=None):
+        check_compressed(value, key)
+    assert not wrongly_compressed, "Some elements were not (un)compressed correctly"
 
     expected = ad.read_zarr(pth)
     assert_equal(adata, expected)
@@ -454,102 +474,6 @@ def test_changed_obs_var_names(tmp_path, diskfmt):
         assert_equal(read, modified, exact=True)
 
 
-@pytest.mark.skipif(not find_spec("loompy"), reason="Loompy is not installed")
-@pytest.mark.parametrize("typ", [np.array, csr_matrix])
-@pytest.mark.parametrize("obsm_mapping", [{}, dict(X_composed=["oanno3", "oanno4"])])
-@pytest.mark.parametrize("varm_mapping", [{}, dict(X_composed2=["vanno3", "vanno4"])])
-def test_readwrite_loom(typ, obsm_mapping, varm_mapping, tmp_path):
-    X = typ(X_list)
-    obs_dim = "meaningful_obs_dim_name"
-    var_dim = "meaningful_var_dim_name"
-    adata_src = ad.AnnData(X, obs=obs_dict, var=var_dict, uns=uns_dict)
-    adata_src.obs_names.name = obs_dim
-    adata_src.var_names.name = var_dim
-    adata_src.obsm["X_a"] = np.zeros((adata_src.n_obs, 2))
-    adata_src.varm["X_b"] = np.zeros((adata_src.n_vars, 3))
-
-    with warnings.catch_warnings():
-        warnings.filterwarnings("ignore", category=NumbaDeprecationWarning)
-        # loompy uses “is” for ints
-        warnings.filterwarnings("ignore", category=SyntaxWarning)
-        warnings.filterwarnings(
-            "ignore",
-            message=r"datetime.datetime.utcnow\(\) is deprecated",
-            category=DeprecationWarning,
-        )
-        adata_src.write_loom(tmp_path / "test.loom", write_obsm_varm=True)
-
-    adata = ad.io.read_loom(
-        tmp_path / "test.loom",
-        sparse=typ is csr_matrix,
-        obsm_mapping=obsm_mapping,
-        obs_names=obs_dim,
-        varm_mapping=varm_mapping,
-        var_names=var_dim,
-        cleanup=True,
-    )
-
-    if isinstance(X, np.ndarray):
-        assert np.allclose(adata.X, X)
-    else:
-        # TODO: this should not be necessary
-        assert np.allclose(adata.X.toarray(), X.toarray())
-    assert "X_a" in adata.obsm_keys()
-    assert adata.obsm["X_a"].shape[1] == 2
-    assert "X_b" in adata.varm_keys()
-    assert adata.varm["X_b"].shape[1] == 3
-    # as we called with `cleanup=True`
-    assert "oanno1b" in adata.uns["loom-obs"]
-    assert "vanno2" in adata.uns["loom-var"]
-    for k, v in obsm_mapping.items():
-        assert k in adata.obsm_keys()
-        assert adata.obsm[k].shape[1] == len(v)
-    for k, v in varm_mapping.items():
-        assert k in adata.varm_keys()
-        assert adata.varm[k].shape[1] == len(v)
-    assert adata.obs_names.name == obs_dim
-    assert adata.var_names.name == var_dim
-
-
-@pytest.mark.skipif(not find_spec("loompy"), reason="Loompy is not installed")
-def test_readloom_deprecations(tmp_path):
-    loom_pth = tmp_path / "test.loom"
-    adata_src = gen_adata((5, 10), obsm_types=[np.ndarray], varm_types=[np.ndarray])
-
-    with warnings.catch_warnings():
-        warnings.filterwarnings("ignore", category=NumbaDeprecationWarning)
-        warnings.filterwarnings(
-            "ignore",
-            message=r"datetime.datetime.utcnow\(\) is deprecated",
-            category=DeprecationWarning,
-        )
-        adata_src.write_loom(loom_pth, write_obsm_varm=True)
-
-    # obsm_names -> obsm_mapping
-    obsm_mapping = {"df": adata_src.obs.columns}
-    with pytest.warns(FutureWarning):
-        depr_result = ad.io.read_loom(loom_pth, obsm_names=obsm_mapping)
-    actual_result = ad.io.read_loom(loom_pth, obsm_mapping=obsm_mapping)
-    assert_equal(actual_result, depr_result)
-    with pytest.raises(ValueError, match=r"ambiguous"), pytest.warns(FutureWarning):
-        ad.io.read_loom(loom_pth, obsm_mapping=obsm_mapping, obsm_names=obsm_mapping)
-
-    # varm_names -> varm_mapping
-    varm_mapping = {"df": adata_src.var.columns}
-    with pytest.warns(FutureWarning):
-        depr_result = ad.io.read_loom(loom_pth, varm_names=varm_mapping)
-    actual_result = ad.io.read_loom(loom_pth, varm_mapping=varm_mapping)
-    assert_equal(actual_result, depr_result)
-    with pytest.raises(ValueError, match=r"ambiguous"), pytest.warns(FutureWarning):
-        ad.io.read_loom(loom_pth, varm_mapping=varm_mapping, varm_names=varm_mapping)
-
-    # positional -> keyword
-    with pytest.warns(FutureWarning, match=r"sparse"):
-        depr_result = ad.io.read_loom(loom_pth, True)  # noqa: FBT003
-    actual_result = ad.io.read_loom(loom_pth, sparse=True)
-    assert type(depr_result.X) == type(actual_result.X)
-
-
 def test_read_csv():
     adata = ad.io.read_csv(HERE / "data" / "adata.csv")
     assert adata.obs_names.tolist() == ["r1", "r2", "r3"]
@@ -572,15 +496,33 @@ def test_read_tsv_iter():
     assert adata.X.tolist() == X_list
 
 
-@pytest.mark.parametrize("typ", [np.array, csr_matrix])
+@pytest.mark.parametrize(
+    "typ",
+    [
+        pytest.param(
+            jnp_array_or_idempotent, id="jax.array", marks=pytest.mark.array_api
+        ),
+        np.array,
+        csr_matrix,
+    ],
+)
 def test_write_csv(typ, tmp_path):
     X = typ(X_list)
     adata = ad.AnnData(X, obs=obs_dict, var=var_dict, uns=uns_dict)
     adata.write_csvs(tmp_path / "test_csv_dir", skip_data=False)
 
 
-@pytest.mark.parametrize("typ", [np.array, csr_matrix])
-def test_write_csv_view(typ, tmp_path):
+@pytest.mark.parametrize(
+    "typ",
+    [
+        pytest.param(
+            jnp_array_or_idempotent, id="jax.array", marks=pytest.mark.array_api
+        ),
+        np.array,
+        csr_matrix,
+    ],
+)
+def test_write_csv_view(typ, tmp_path: Path) -> None:
     # https://github.com/scverse/anndata/issues/401
     import hashlib
 
@@ -615,19 +557,18 @@ def test_write_csv_view(typ, tmp_path):
     ("read", "write", "name"),
     [
         pytest.param(ad.read_h5ad, ad.io.write_h5ad, "test_empty.h5ad"),
-        pytest.param(
-            ad.io.read_loom,
-            ad.io.write_loom,
-            "test_empty.loom",
-            marks=pytest.mark.xfail(reason="Loom can’t handle 0×0 matrices"),
-        ),
         pytest.param(ad.read_zarr, ad.io.write_zarr, "test_empty.zarr"),
     ],
 )
-def test_readwrite_empty(read, write, name, tmp_path):
-    adata = ad.AnnData(uns=dict(empty=np.array([], dtype=float)))
+@pytest.mark.parametrize(
+    "xp_array",
+    [np.array, pytest.param(jnp_array_or_idempotent, marks=pytest.mark.array_api)],
+)
+def test_readwrite_empty(read, write, name: str, tmp_path: Path, xp_array) -> None:
+    adata = ad.AnnData(uns=dict(empty=xp_array([]).astype(float)))
     write(tmp_path / name, adata)
     ad_read = read(tmp_path / name)
+    assert ad_read.uns["empty"] is not None
     assert ad_read.uns["empty"].shape == (0,)
 
 
@@ -655,22 +596,22 @@ def test_read_umi_tools():
 def test_write_categorical(
     *, tmp_path: Path, diskfmt: Literal["h5ad", "zarr"], s2c: bool
 ) -> None:
+    adata_pth = tmp_path / f"adata.{diskfmt}"
+    obs = dict(
+        str=pd.array(["a", "a", "b", pd.NA, pd.NA], dtype="string"),
+        cat=pd.Categorical(["a", "a", "b", np.nan, np.nan]),
+        **(dict(obj=["a", "a", "b", np.nan, np.nan]) if s2c else {}),
+    )
+    orig = ad.AnnData(obs=pd.DataFrame(obs))
     with ad.settings.override(allow_write_nullable_strings=True):
-        adata_pth = tmp_path / f"adata.{diskfmt}"
-        obs = dict(
-            str=pd.array(["a", "a", "b", pd.NA, pd.NA], dtype="string"),
-            cat=pd.Categorical(["a", "a", "b", np.nan, np.nan]),
-            **(dict(obj=["a", "a", "b", np.nan, np.nan]) if s2c else {}),
-        )
-        orig = ad.AnnData(obs=pd.DataFrame(obs))
         getattr(orig, f"write_{diskfmt}")(
             adata_pth, convert_strings_to_categoricals=s2c
         )
-        curr: ad.AnnData = getattr(ad, f"read_{diskfmt}")(adata_pth)
-        assert np.all(orig.obs.notna() == curr.obs.notna())
-        assert np.all(orig.obs.stack().dropna() == curr.obs.stack().dropna())
-        assert curr.obs["str"].dtype == ("category" if s2c else "string")
-        assert curr.obs["cat"].dtype == "category"
+    curr: ad.AnnData = getattr(ad, f"read_{diskfmt}")(adata_pth)
+    assert np.all(orig.obs.notna() == curr.obs.notna())
+    assert np.all(orig.obs.stack().dropna() == curr.obs.stack().dropna())
+    assert curr.obs["str"].dtype == ("category" if s2c else "string")
+    assert curr.obs["cat"].dtype == "category"
 
 
 def test_write_categorical_index(tmp_path, diskfmt):
@@ -806,8 +747,15 @@ def _do_roundtrip(
 
 
 @pytest.fixture
-def roundtrip(diskfmt):
+def roundtrip(
+    diskfmt: Literal["h5ad", "zarr"],
+) -> Callable[[ad.AnnData, Path], ad.AnnData]:
     return partial(_do_roundtrip, diskfmt=diskfmt)
+
+
+@pytest.fixture
+def roundtrip2(diskfmt2):
+    return partial(_do_roundtrip, diskfmt=diskfmt2)
 
 
 def test_write_string_types(tmp_path, diskfmt, roundtrip):
@@ -828,67 +776,61 @@ def test_write_string_types(tmp_path, diskfmt, roundtrip):
 
 
 @pytest.mark.skipif(not find_spec("scanpy"), reason="Scanpy is not installed")
-def test_scanpy_pbmc68k(tmp_path, diskfmt, roundtrip, diskfmt2):
-    roundtrip2 = partial(_do_roundtrip, diskfmt=diskfmt2)
+def test_scanpy_pbmc68k(tmp_path, diskfmt, roundtrip, diskfmt2, roundtrip2):
+    import scanpy as sc
 
-    filepth1 = tmp_path / f"test1.{diskfmt}"
-    filepth2 = tmp_path / f"test2.{diskfmt2}"
-
-    with warnings.catch_warnings():
-        warnings.filterwarnings(
-            "ignore", message=r"Importing read_.* from `anndata` is deprecated"
-        )
-        import scanpy as sc
-
+    # TODO: remove filters (here and elsewhere) once https://github.com/scverse/scanpy/issues/3879 is fixed
     with warnings.catch_warnings():
         warnings.simplefilter("ignore", ad.OldFormatWarning)
+        warnings.filterwarnings(
+            "ignore", r"Moving element.*uns.*to.*obsp", FutureWarning
+        )
         pbmc = sc.datasets.pbmc68k_reduced()
 
-    from_disk1 = roundtrip(pbmc, filepth1)  # Do we read okay
-    from_disk2 = roundtrip2(from_disk1, filepth2)  # Can we round trip
+    # Do we read okay
+    from_disk1 = roundtrip(pbmc, tmp_path / f"test1.{diskfmt}")
+    # Can we round trip
+    from_disk2 = roundtrip2(from_disk1, tmp_path / f"test2.{diskfmt2}")
 
     assert_equal(pbmc, from_disk1)  # Not expected to be exact due to `nan`s
     assert_equal(pbmc, from_disk2)
 
 
+@pytest.mark.filterwarnings(r"ignore:Observation names are not unique:UserWarning")
 @pytest.mark.skipif(not find_spec("scanpy"), reason="Scanpy is not installed")
-def test_scanpy_krumsiek11(tmp_path, diskfmt, roundtrip):
-    filepth = tmp_path / f"test.{diskfmt}"
-    with warnings.catch_warnings():
-        warnings.filterwarnings(
-            "ignore", message=r"Importing read_.* from `anndata` is deprecated"
-        )
-        import scanpy as sc
+def test_scanpy_krumsiek11(
+    tmp_path: Path,
+    diskfmt: Literal["h5ad", "zarr"],
+    roundtrip: Callable[[ad.AnnData, Path], ad.AnnData],
+) -> None:
+    import scanpy as sc
 
-    # TODO: this should be fixed in scanpy instead
-    with pytest.warns(UserWarning, match=r"Observation names are not unique"):  # noqa: PT031
+    with warnings.catch_warnings():
         warnings.filterwarnings(
             "ignore", r".*first_column_names.*no longer positional", FutureWarning
         )
         orig = sc.datasets.krumsiek11()
     del orig.uns["highlights"]  # Can’t write int keys
-    # Can’t write "string" dtype: https://github.com/scverse/anndata/issues/679
-    orig.obs["cell_type"] = orig.obs["cell_type"].astype(str)
-    with pytest.warns(UserWarning, match=r"Observation names are not unique"):
-        curr = roundtrip(orig, filepth)
-
+    # Depending on pd.options.future.infer_string, this becomes either `object` or `'string'`
+    orig.var.columns = orig.var.columns.astype(str)
+    with ad.settings.override(allow_write_nullable_strings=True):
+        curr = roundtrip(orig, tmp_path / f"test.{diskfmt}")
+    # These categories are constructed manually in scanpy's code so are not "roundtripped" from disk.
+    orig.obs["cell_type"] = orig.obs["cell_type"].astype(curr.obs["cell_type"].dtype)
     assert_equal(orig, curr, exact=True)
 
 
 # Checking if we can read legacy zarr files
 # TODO: Check how I should add this file to the repo
 @pytest.mark.filterwarnings("ignore::anndata.OldFormatWarning")
+@pytest.mark.filterwarnings(r"ignore:Moving element.*uns.*to.*obsp:FutureWarning")
 @pytest.mark.skipif(not find_spec("scanpy"), reason="Scanpy is not installed")
 @pytest.mark.skipif(
     not Path(HERE / "data/pbmc68k_reduced_legacy.zarr.zip").is_file(),
     reason="File not present.",
 )
-def test_backwards_compat_zarr():
-    with warnings.catch_warnings():
-        warnings.filterwarnings(
-            "ignore", message=r"Importing read_.* from `anndata` is deprecated"
-        )
-        import scanpy as sc
+def test_backwards_compat_zarr() -> None:
+    import scanpy as sc
     import zarr
 
     pbmc_orig = sc.datasets.pbmc68k_reduced()
@@ -929,11 +871,16 @@ def test_adata_in_uns(tmp_path, diskfmt, roundtrip):
     [
         pytest.param(dict(base=None), id="dict_val"),
         pytest.param(
-            pd.DataFrame(dict(col_0=["string", None])).convert_dtypes(), id="df"
+            lambda: pd.DataFrame(
+                dict(col_0=pd.array(["string", None], dtype="string"))
+            ),
+            id="df",
         ),
     ],
 )
 def test_none_dict_value_in_uns(diskfmt, tmp_path, roundtrip, uns_val):
+    if callable(uns_val):
+        uns_val = uns_val()
     pth = tmp_path / f"adata_dtype.{diskfmt}"
 
     orig = ad.AnnData(np.ones((3, 4)), uns=dict(val=uns_val))
@@ -967,13 +914,28 @@ def test_h5py_attr_limit(tmp_path):
 @pytest.mark.parametrize(
     "elem_key", ["obs", "var", "obsm", "varm", "layers", "obsp", "varp", "uns"]
 )
-def test_forward_slash_key(elem_key, tmp_path):
+@pytest.mark.parametrize("store_type", ["zarr", "h5ad"])
+@pytest.mark.parametrize("disallow_forward_slash_in_h5ad", [True, False])
+def test_forward_slash_key(
+    elem_key: Literal["obs", "var", "obsm", "varm", "layers", "obsp", "varp", "uns"],
+    tmp_path: Path,
+    store_type: Literal["zarr", "h5ad"],
+    *,
+    disallow_forward_slash_in_h5ad: bool,
+):
     a = ad.AnnData(np.ones((10, 10)))
     getattr(a, elem_key)["bad/key"] = np.ones(
         (10,) if elem_key in ["obs", "var"] else (10, 10)
     )
-    with pytest.raises(ValueError, match="Forward slashes"):
-        a.write_h5ad(tmp_path / "does_not_matter_the_path.h5ad")
+    with (
+        ad.settings.override(
+            disallow_forward_slash_in_h5ad=disallow_forward_slash_in_h5ad
+        ),
+        pytest.raises(ValueError, match=r"Forward slashes")
+        if store_type == "zarr" or disallow_forward_slash_in_h5ad
+        else pytest.warns(FutureWarning, match=r"Forward slashes"),
+    ):
+        getattr(a, f"write_{store_type}")(tmp_path / "does_not_matter_the_path.h5ad")
 
 
 @pytest.mark.skipif(
@@ -996,11 +958,7 @@ def test_read_lazy_import_error(func, tmp_path):
 @pytest.mark.zarr_io
 def test_write_elem_consolidated(tmp_path: Path):
     ad.AnnData(np.ones((10, 10))).write_zarr(tmp_path)
-    g = (
-        zarr.convenience.open_consolidated(tmp_path)
-        if is_zarr_v2()
-        else zarr.open(tmp_path)
-    )
+    g = zarr.open(tmp_path)
     with pytest.raises(
         ValueError, match="Cannot overwrite/edit a store with consolidated metadata"
     ):
@@ -1008,7 +966,6 @@ def test_write_elem_consolidated(tmp_path: Path):
 
 
 @pytest.mark.zarr_io
-@pytest.mark.skipif(is_zarr_v2(), reason="zarr v3 package test")
 def test_write_elem_version_mismatch(tmp_path: Path):
     zarr_path = tmp_path / "foo.zarr"
     adata = ad.AnnData(np.ones((10, 10)))
