@@ -1,6 +1,7 @@
 from __future__ import annotations
 
 import string
+from typing import cast
 
 import numpy as np
 import pandas as pd
@@ -31,6 +32,22 @@ def test_columns(df, dataset2d):
     assert np.all(dataset2d.columns.sort_values() == df.columns.sort_values())
 
 
+@pytest.mark.parametrize("same_columns", [True, False], ids=["same", "different"])
+def test_columns_setter(df, dataset2d: Dataset2D, *, same_columns: bool):
+    dataset2d_orig = dataset2d.copy()
+    with (
+        pytest.warns(
+            UserWarning, match=r"Renaming or reordering columns on `Dataset2D`"
+        )
+        if same_columns
+        else pytest.raises(ValueError, match=r"Trying to rename the keys")
+    ):
+        dataset2d.columns = (
+            dataset2d.columns if same_columns else pd.Index(["not", "a", "column"])
+        )
+    assert dataset2d.equals(dataset2d_orig)
+
+
 def test_to_memory(df, dataset2d):
     memory_df = dataset2d.to_memory()
     assert np.all(df == memory_df)
@@ -42,6 +59,8 @@ def test_getitem(df, dataset2d):
     col = df.columns[0]
     assert np.all(dataset2d[col] == df[col])
 
+
+def test_getitem_empty(df, dataset2d):
     empty_dset = dataset2d[[]]
     assert empty_dset.shape == (df.shape[0], 0)
     assert np.all(empty_dset.index == dataset2d.index)
@@ -57,40 +76,81 @@ def test_backed_property(dataset2d):
     assert not dataset2d.is_backed
 
 
-def test_index_dim(dataset2d):
-    assert dataset2d.index_dim == "index"
-    assert dataset2d.true_index_dim == dataset2d.index_dim
-
-    col = next(iter(dataset2d.keys()))
+def test_true_index_dim_column_subset(dataset2d, df):
+    col_iter = iter(dataset2d.keys())
+    col = next(col_iter)
     dataset2d.true_index_dim = col
-    assert dataset2d.index_dim == "index"
-    assert dataset2d.true_index_dim == col
 
-    with pytest.raises(ValueError, match=r"Unknown variable `test`\."):
-        dataset2d.true_index_dim = "test"
+    # Ensure we can actually select columns properly that are not the index column
+    cols = [next(col_iter), next(col_iter)]
+    df_expected = dataset2d[cols].to_memory()
+    # account for the fact that we manually set `true_index_dim`
+    df.index = df[col]
+    df.index.name = None
+    pd.testing.assert_frame_equal(df_expected, df[cols])
 
-    dataset2d.true_index_dim = None
-    assert dataset2d.true_index_dim == dataset2d.index_dim
+
+def true_index_set_from_existing_col(dataset2D: Dataset2D) -> None:
+    col = cast("str", next(iter(dataset2D.keys())))
+    dataset2D.true_index_dim = col
+    assert dataset2D.index_dim == "index"
+    assert dataset2D.true_index_dim == col
 
 
-def test_index(dataset2d):
-    alphabet = np.asarray(
-        list(string.ascii_letters + string.digits + string.punctuation)
+def true_index_set_from_unknown_col(dataset2D: Dataset2D) -> None:
+    with (
+        pytest.raises(ValueError, match=r"Unknown variable `test`\."),
+    ):
+        dataset2D.true_index_dim = "test"
+
+
+def true_index_set_unset(dataset2D: Dataset2D) -> None:
+    dataset2D.true_index_dim = None
+    assert dataset2D.true_index_dim == dataset2D.index_dim
+
+
+def test_index_setting_from_existing_column(dataset2d: Dataset2D) -> None:
+    col = cast("str", next(iter(dataset2d.keys())))
+    old_index_dim = dataset2d.index_dim
+
+    dataset2d.index = dataset2d[col]
+
+    assert dataset2d.index_dim == col
+    assert old_index_dim not in (
+        *dataset2d.columns,
+        dataset2d.index_dim,
+        dataset2d.index.name,
     )
-    new_idx = pd.Index(
-        [
-            "".join(np.random.choice(alphabet, size=10))
-            for _ in range(dataset2d.shape[0])
-        ],
-        name="test_index",
+
+
+def test_index_setting_from_different_true_index_overrides_index(
+    dataset2d: Dataset2D,
+) -> None:
+    col = cast("str", next(iter(dataset2d.keys())))
+    dataset2d.true_index_dim = col
+    old_index_dim = dataset2d.index_dim
+    dataset2d.index = dataset2d[dataset2d.true_index_dim]
+    assert dataset2d.index_dim == dataset2d.true_index_dim
+    assert old_index_dim not in (
+        *dataset2d.columns,
+        dataset2d.index_dim,
+        dataset2d.index.name,
     )
+
+
+def test_set_index_from_outside_dataset(dataset2d: Dataset2D) -> None:
+    alphabet = np.asarray([*string.ascii_letters, *string.digits, *string.punctuation])
+    new_idx = set()
+    while len(new_idx) < dataset2d.shape[0]:  # if we get duplicates, we just try again
+        new_idx.add("".join(np.random.choice(alphabet, size=10)))
+    new_idx = pd.Index(new_idx, name="test_index")
 
     col = next(iter(dataset2d.keys()))
     dataset2d.true_index_dim = col
 
     dataset2d.index = new_idx
     assert np.all(dataset2d.index == new_idx)
-    assert dataset2d.true_index_dim == dataset2d.index_dim == new_idx.name
+    assert dataset2d.index_dim == new_idx.name
     assert list(dataset2d.ds.coords.keys()) == [new_idx.name]
 
 
@@ -256,7 +316,7 @@ def test_dataset_2d_set_with_bad_obj(da, pattern, dataset_2d_one_column):
 def test_dataset_2d_set_index(data, dataset_2d_one_column):
     with pytest.raises(
         KeyError,
-        match="Cannot set obs_names as a variable",
+        match="Cannot set the index dimension obs_names",
     ):
         dataset_2d_one_column["obs_names"] = data
 
